refactoring component structure
