+++ bioinformatics-software-api/workflow/tasks/blast.js
@@ -1,3727 +1,13 @@
 /**
  * Created by anuradhawick on 9/11/17.
  */
-import uuidv4 from 'uuid/v4';
-import fs from 'fs';
-import path from 'path';
-
 import WrapBlast from '../../python-wrappers/blast/wrap-blast';
 
 
 export default class Blast {
 
-    static data = `BLASTN 2.7.0+
-Reference: Stephen F. Altschul, Thomas L. Madden, Alejandro
-A. Schaffer, Jinghui Zhang, Zheng Zhang, Webb Miller, and
-David J. Lipman (1997), "Gapped BLAST and PSI-BLAST: a new
-generation of protein database search programs", Nucleic
-Acids Res. 25:3389-3402.
-
-
-RID: Y9HYVUKF015
-
-
-Database: Nucleotide collection (nt)
-           44,928,843 sequences; 160,111,102,915 total letters
-Query= HSBGPG Human gene for bone gla protein (BGP)
-
-Length=1231
-
-
-                                                                   Score     E
-Sequences producing significant alignments:                       (Bits)  Value
-
-NG_047015.1  Homo sapiens bone gamma-carboxyglutamate protein ...  2221    0.0   
-AL135927.14  Human DNA sequence from clone RP11-54H19 on chrom...  2221    0.0   
-DQ007079.1  Homo sapiens bone gamma-carboxyglutamate (gla) pro...  2221    0.0   
-X04143.1  Human gene for bone gla protein (BGP)                    2221    0.0   
-AC007227.3  Homo sapiens chromosome 1 clone RPCI-11_54H19, com...  2221    0.0   
-AC193856.2  Pan troglodytes BAC clone CH251-490M23 from chromo...  2197    0.0   
-AC193547.5  Rhesus Macaque BAC CH250-5J10 () complete sequence     1952    0.0   
-LT160000.1  Macaca fascicularis complete genome, chromosome chr1   1930    0.0   
-DQ977353.1  Pan troglodytes BGLAP (BGLAP) gene, complete cds       1896    0.0   
-DQ976476.1  Gorilla gorilla BGLAP (BGLAP) gene, complete cds       1878    0.0   
-DQ977503.1  Pongo pygmaeus BGLAP (BGLAP) gene, complete cds        1838    0.0   
-DQ977098.1  Macaca nemestrina BGLAP (BGLAP) gene, complete cds     1712    0.0   
-DQ976615.1  Ateles geoffroyi BGLAP (BGLAP) gene, complete cds      1642    0.0   
-DQ976876.1  Macaca mulatta BGLAP (BGLAP) gene, complete cds        1626    0.0   
-DQ977209.1  Pan paniscus BGLAP (BGLAP) gene, partial cds           1409    0.0   
-DQ976754.1  Lagothrix lagotricha BGLAP (BGLAP) gene, partial cds   758     0.0   
-DQ976983.1  Saguinus labiatus BGLAP (BGLAP) gene, partial cds      724     0.0   
-XM_020169839.1  PREDICTED: Castor canadensis polyamine-modulat...  517     4e-142
-NG_053096.1  Homo sapiens BGLAP promoter region (LOC110013312)...  515     1e-141
-NM_199173.5  Homo sapiens bone gamma-carboxyglutamate protein ...  508     2e-139
-NM_001199662.1  Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLA...  508     2e-139
-NM_001199664.1  Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLA...  508     2e-139
-NM_001199663.1  Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLA...  508     2e-139
-NM_001199661.1  Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLA...  508     2e-139
-BC033656.2  Homo sapiens polyamine-modulated factor 1, mRNA (c...  508     2e-139
-XM_016928883.1  PREDICTED: Pan troglodytes bone gamma-carboxyg...  502     9e-138
-XM_004026955.2  PREDICTED: Gorilla gorilla gorilla bone gamma-...  495     1e-135
-XM_002810036.3  PREDICTED: Pongo abelii bone gamma-carboxyglut...  493     5e-135
-XM_008974558.2  PREDICTED: Pan paniscus bone gamma-carboxyglut...  477     4e-130
-X51699.1  Human mRNA for bone Gla protein                          471     2e-128
-XM_003259455.1  PREDICTED: Nomascus leucogenys bone gamma-carb...  466     7e-127
-XM_005541474.2  PREDICTED: Macaca fascicularis bone gamma-carb...  450     5e-122
-XM_011769795.1  PREDICTED: Macaca nemestrina bone gamma-carbox...  450     5e-122
-XM_011992748.1  PREDICTED: Mandrillus leucophaeus bone gamma-c...  450     5e-122
-XM_012078677.1  PREDICTED: Cercocebus atys bone gamma-carboxyg...  441     3e-119
-XM_010366604.1  PREDICTED: Rhinopithecus roxellana osteocalcin...  441     3e-119
-XM_011928386.1  PREDICTED: Colobus angolensis palliatus bone g...  435     1e-117
-XM_017852456.1  PREDICTED: Rhinopithecus bieti bone gamma-carb...  430     5e-116
-XR_001881540.1  PREDICTED: Rhinopithecus bieti polyamine-modul...  430     5e-116
-XM_007976754.1  PREDICTED: Chlorocebus sabaeus bone gamma-carb...  430     5e-116
-XM_003892817.4  PREDICTED: Papio anubis bone gamma-carboxyglut...  414     4e-111
-XM_012449840.2  PREDICTED: Aotus nancymaae bone gamma-carboxyg...  403     7e-108
-BC113432.1  Homo sapiens bone gamma-carboxyglutamate (gla) pro...  399     8e-107
-XM_010348451.1  PREDICTED: Saimiri boliviensis boliviensis bon...  392     1e-104
-XM_010348450.1  PREDICTED: Saimiri boliviensis boliviensis bon...  392     1e-104
-XM_003937835.2  PREDICTED: Saimiri boliviensis boliviensis bon...  392     1e-104
-XM_008984594.2  PREDICTED: Callithrix jacchus bone gamma-carbo...  383     6e-102
-XM_017510005.1  PREDICTED: Cebus capucinus imitator bone gamma...  369     1e-97 
-XM_010366603.1  PREDICTED: Rhinopithecus roxellana osteocalcin...  343     6e-90 
-XM_012658608.1  PREDICTED: Propithecus coquereli bone gamma-ca...  293     9e-75 
-
-ALIGNMENTS
->NG_047015.1 Homo sapiens bone gamma-carboxyglutamate protein (BGLAP), RefSeqGene 
-on chromosome 1
-Length=8173
-
- Score = 2221 bits (2462),  Expect = 0.0
- Identities = 1231/1231 (100%), Gaps = 0/1231 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4923  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  4982
-
-Query  61    CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4983  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  5042
-
-Query  121   GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5043  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  5102
-
-Query  181   GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  240
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5103  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  5162
-
-Query  241   CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  300
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5163  CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  5222
-
-Query  301   GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  360
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5223  GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  5282
-
-Query  361   CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  420
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5283  CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  5342
-
-Query  421   TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  480
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5343  TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  5402
-
-Query  481   CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  540
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5403  CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  5462
-
-Query  541   CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  600
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5463  CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  5522
-
-Query  601   GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  660
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5523  GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  5582
-
-Query  661   CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  720
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5583  CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  5642
-
-Query  721   ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  780
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5643  ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  5702
-
-Query  781   CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  840
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5703  CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  5762
-
-Query  841   GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  900
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5763  GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  5822
-
-Query  901   AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  960
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5823  AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  5882
-
-Query  961   CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  1020
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5883  CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  5942
-
-Query  1021  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  1080
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5943  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  6002
-
-Query  1081  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  1140
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  6003  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  6062
-
-Query  1141  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  1200
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  6063  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  6122
-
-Query  1201  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||
-Sbjct  6123  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  6153
-
-
->AL135927.14 Human DNA sequence from clone RP11-54H19 on chromosome 1, complete 
-sequence
-Length=164168
-
- Score = 2221 bits (2462),  Expect = 0.0
- Identities = 1231/1231 (100%), Gaps = 0/1231 (0%)
- Strand=Plus/Plus
-
-Query  1       GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  129770  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  129829
-
-Query  61      CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  129830  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  129889
-
-Query  121     GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  129890  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  129949
-
-Query  181     GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  240
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  129950  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  130009
-
-Query  241     CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  300
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130010  CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  130069
-
-Query  301     GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  360
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130070  GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  130129
-
-Query  361     CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  420
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130130  CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  130189
-
-Query  421     TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  480
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130190  TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  130249
-
-Query  481     CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  540
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130250  CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  130309
-
-Query  541     CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  600
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130310  CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  130369
-
-Query  601     GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  660
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130370  GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  130429
-
-Query  661     CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  720
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130430  CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  130489
-
-Query  721     ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  780
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130490  ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  130549
-
-Query  781     CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  840
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130550  CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  130609
-
-Query  841     GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  900
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130610  GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  130669
-
-Query  901     AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  960
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130670  AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  130729
-
-Query  961     CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  1020
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130730  CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  130789
-
-Query  1021    CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  1080
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130790  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  130849
-
-Query  1081    GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  1140
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130850  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  130909
-
-Query  1141    TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  1200
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  130910  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  130969
-
-Query  1201    ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-               |||||||||||||||||||||||||||||||
-Sbjct  130970  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  131000
-
-
->DQ007079.1 Homo sapiens bone gamma-carboxyglutamate (gla) protein (osteocalcin) 
-(BGLAP) gene, complete cds
-Length=10147
-
- Score = 2221 bits (2462),  Expect = 0.0
- Identities = 1231/1231 (100%), Gaps = 0/1231 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  3971  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  4030
-
-Query  61    CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4031  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  4090
-
-Query  121   GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4091  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  4150
-
-Query  181   GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  240
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4151  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  4210
-
-Query  241   CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  300
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4211  CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  4270
-
-Query  301   GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  360
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4271  GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  4330
-
-Query  361   CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  420
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4331  CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  4390
-
-Query  421   TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  480
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4391  TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  4450
-
-Query  481   CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  540
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4451  CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  4510
-
-Query  541   CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  600
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4511  CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  4570
-
-Query  601   GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  660
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4571  GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  4630
-
-Query  661   CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  720
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4631  CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  4690
-
-Query  721   ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  780
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4691  ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  4750
-
-Query  781   CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  840
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4751  CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  4810
-
-Query  841   GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  900
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4811  GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  4870
-
-Query  901   AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  960
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4871  AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  4930
-
-Query  961   CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  1020
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4931  CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  4990
-
-Query  1021  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  1080
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  4991  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  5050
-
-Query  1081  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  1140
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5051  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  5110
-
-Query  1141  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  1200
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  5111  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  5170
-
-Query  1201  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||
-Sbjct  5171  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  5201
-
-
->X04143.1 Human gene for bone gla protein (BGP)
-Length=1675
-
- Score = 2221 bits (2462),  Expect = 0.0
- Identities = 1231/1231 (100%), Gaps = 0/1231 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  346   GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  405
-
-Query  61    CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  406   CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  465
-
-Query  121   GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  466   GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  525
-
-Query  181   GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  240
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  526   GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  585
-
-Query  241   CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  300
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  586   CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  645
-
-Query  301   GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  360
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  646   GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  705
-
-Query  361   CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  420
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  706   CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  765
-
-Query  421   TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  480
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  766   TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  825
-
-Query  481   CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  540
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  826   CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  885
-
-Query  541   CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  600
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  886   CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  945
-
-Query  601   GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  660
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  946   GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  1005
-
-Query  661   CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  720
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1006  CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  1065
-
-Query  721   ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  780
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1066  ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  1125
-
-Query  781   CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  840
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1126  CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  1185
-
-Query  841   GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  900
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1186  GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  1245
-
-Query  901   AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  960
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1246  AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  1305
-
-Query  961   CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  1020
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1306  CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  1365
-
-Query  1021  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  1080
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1366  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  1425
-
-Query  1081  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  1140
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1426  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  1485
-
-Query  1141  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  1200
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1486  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  1545
-
-Query  1201  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||
-Sbjct  1546  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1576
-
-
->AC007227.3 Homo sapiens chromosome 1 clone RPCI-11_54H19, complete sequence
-Length=164179
-
- Score = 2221 bits (2462),  Expect = 0.0
- Identities = 1231/1231 (100%), Gaps = 0/1231 (0%)
- Strand=Plus/Minus
-
-Query  1      GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  34399  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  34340
-
-Query  61     CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  34339  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  34280
-
-Query  121    GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  34279  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  34220
-
-Query  181    GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  240
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  34219  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  34160
-
-Query  241    CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  300
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  34159  CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  34100
-
-Query  301    GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  360
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  34099  GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  34040
-
-Query  361    CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  420
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  34039  CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  33980
-
-Query  421    TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  480
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33979  TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  33920
-
-Query  481    CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  540
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33919  CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  33860
-
-Query  541    CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  600
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33859  CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  33800
-
-Query  601    GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  660
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33799  GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  33740
-
-Query  661    CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  720
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33739  CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  33680
-
-Query  721    ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  780
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33679  ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  33620
-
-Query  781    CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  840
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33619  CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  33560
-
-Query  841    GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  900
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33559  GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  33500
-
-Query  901    AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  960
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33499  AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  33440
-
-Query  961    CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  1020
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33439  CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  33380
-
-Query  1021   CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  1080
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33379  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  33320
-
-Query  1081   GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  1140
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33319  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  33260
-
-Query  1141   TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  1200
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  33259  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  33200
-
-Query  1201   ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-              |||||||||||||||||||||||||||||||
-Sbjct  33199  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  33169
-
-
->AC193856.2 Pan troglodytes BAC clone CH251-490M23 from chromosome 1, complete 
-sequence
-Length=186904
-
- Score = 2197 bits (2436),  Expect = 0.0
- Identities = 1226/1231 (99%), Gaps = 0/1231 (0%)
- Strand=Plus/Minus
-
-Query  1       GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  140210  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  140151
-
-Query  61      CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  140150  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  140091
-
-Query  121     GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  140090  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  140031
-
-Query  181     GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  240
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  140030  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  139971
-
-Query  241     CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  300
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139970  CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTG  139911
-
-Query  301     GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  360
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139910  GCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCT  139851
-
-Query  361     CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGA  420
-               ||||||||||||||||||||||||||||||||||||||||||||| ||||||||||||||
-Sbjct  139850  CTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACACAGGGGAAGGAGGA  139791
-
-Query  421     TGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  480
-               |||||||||||||||||||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  139790  TGAGGGCCCTGGGGATGAGCTGGGGTGAGCCAGGCTCCCTTTCCTTTGCAGGTGCGAAGC  139731
-
-Query  481     CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  540
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139730  CCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGA  139671
-
-Query  541     CCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  600
-               |||||||||||||||||| |||||||||||||||||||||||||||||||||||||||||
-Sbjct  139670  CCCTCCCCTCTCACCCTGTTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTG  139611
-
-Query  601     GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  660
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139610  GCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATG  139551
-
-Query  661     CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  720
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139550  CCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAG  139491
-
-Query  721     ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  780
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139490  ACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGC  139431
-
-Query  781     CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  840
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139430  CCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGT  139371
-
-Query  841     GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  900
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139370  GGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGG  139311
-
-Query  901     AGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  960
-               ||||||||||||||||||| ||||||||||||||||||||||||||||||||||||||||
-Sbjct  139310  AGGGATGGGCATTTTGCACAGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGT  139251
-
-Query  961     CCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  1020
-               ||||||||||||| ||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139250  CCCCTACCCGGATACCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGA  139191
-
-Query  1021    CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  1080
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139190  CGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTA  139131
-
-Query  1081    GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  1140
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139130  GGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTC  139071
-
-Query  1141    TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  1200
-               ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  139070  TTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCATC  139011
-
-Query  1201    ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-               |||||||||||||||||||||||||||||||
-Sbjct  139010  ATCCCAGCTGCTCCCAAATAAACTCCAGAAG  138980
-
-
->AC193547.5 Rhesus Macaque BAC CH250-5J10 () complete sequence
-Length=175357
-
- Score = 1952 bits (2164),  Expect = 0.0
- Identities = 1174/1232 (95%), Gaps = 6/1232 (0%)
- Strand=Plus/Plus
-
-Query  1      GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-              ||||||||||||||||||||||| |||||||||||||||| ||||||||||||| | |||
-Sbjct  79296  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGGCACGCCCCTCCTCATCACCGGG  79355
-
-Query  61     -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-               |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  79356  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  79415
-
-Query  120    AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-              ||||| |||||| | ||||||| ||||| ||||||||||||||||||||||||||||  |
-Sbjct  79416  AGCAGGAGCCCAACACAGCCACTGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCT  79475
-
-Query  180    GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGG  239
-              |||||||||| ||||||||||| |||||||||||||||||||||||||||||||||||||
-Sbjct  79476  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGG  79535
-
-Query  240    CCGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTT  299
-              |||||  ||||||||||||||||  |||||||||||||||||||||||||||||||||||
-Sbjct  79536  CCGCACAGCAGTGGGGGCTGAGAAAAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTT  79595
-
-Query  300    GGCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGC  359
-              ||||||||||||||||| ||||||| |||| |||||||||||||||||||||||||||||
-Sbjct  79596  GGCTGGCAGTCCCTTTGTAGTCTAATCACCCTGTTGCAGGCTCAATCCATTTGCCCCAGC  79655
-
-Query  360    TCTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGG  419
-              ||||||||||||||||||||||| |||||||||||||||||||||| |||||||||||||
-Sbjct  79656  TCTGCCCTTGCAGAGGGAGAGGAAGGAAGAGCAAGCTGCCCGAGACACAGGGGAAGGAGG  79715
-
-Query  420    ATGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAG  479
-              ||||| ||||||||||||||| ||||||||||||||||||||||||||||||||||||| 
-Sbjct  79716  ATGAGAGCCCTGGGGATGAGCAGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAA  79775
-
-Query  480    CCCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGG  539
-              |||||||||||||||||||||||||||||||||| |||||||||||||||||| ||||||
-Sbjct  79776  CCCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTAAGAGGATGGACCTGATGGGCTCCTGG  79835
-
-Query  540    ACCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCT  599
-              |||||    ||||||| |||||||||||||||||||||| ||||||||||||||||||||
-Sbjct  79836  ACCCT----TCTCACCTTGGTCCCTCAGTCTCATTCCCCTACTCCTGCCACCTCCTGTCT  79891
-
-Query  600    GGCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGAT  659
-              |||  ||||||||||||||||||| || ||||||||||||||||||||||||||||||||
-Sbjct  79892  GGCTGTCAGGAAGGCCAGCCTGCTGCCAACCTGATCCTCCCAAACCCAGAGCCACCTGAT  79951
-
-Query  660    GCCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGA  719
-               ||||||||||||||||||||||||||||||||||||||||||||||| ||| |||||||
-Sbjct  79952  TCCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGCGGTGGTGAAGA  80011
-
-Query  720    GACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGG  779
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  80012  GACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGG  80071
-
-Query  780    CCCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTG  839
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||| ||
-Sbjct  80072  CCCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGCTG  80131
-
-Query  840    TGGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAG  899
-              ||||||||| |||||||||| ||||||||||| ||||||||||||||||||||| |||||
-Sbjct  80132  TGGTGGGGGAACAGGCAGCCAGCCCTGGTGGGGACCCTGGAGCCCCATGTGTAGTGAGAG  80191
-
-Query  900    GAGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAG  959
-              ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  80192  GAGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAG  80251
-
-Query  960    TCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTG  1019
-               ||||||||| |||||||||||||||| ||||||||||||||||||||||||||||||||
-Sbjct  80252  CCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATCCGGACTGTG  80311
-
-Query  1020   ACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCT  1079
-              ||||| ||||||||||||||||||| ||||||||||||||||||||||||||||||||||
-Sbjct  80312  ACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCT  80371
-
-Query  1080   AGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTT  1139
-              |||||||||||| ||||||||||||||||| ||||||||||||||||||||||| |||||
-Sbjct  80372  AGGGTGTCGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCAGGCCCCCTT  80431
-
-Query  1140   CTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCAT  1199
-              |||||| |||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  80432  CTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCCCAT  80491
-
-Query  1200   CATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-              ||| ||||||||||||||||||||| || |||
-Sbjct  80492  CAT-CCAGCTGCTCCCAAATAAACTTCAAAAG  80522
-
-
->LT160000.1 Macaca fascicularis complete genome, chromosome chr1
-Length=229594237
-
- Score = 1930 bits (2140),  Expect = 0.0
- Identities = 1171/1235 (95%), Gaps = 5/1235 (0%)
- Strand=Plus/Plus
-
-Query  1          GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-                  ||||||||||||||||||||||| |||||||||||||||| ||||||||||||| | |||
-Sbjct  135568678  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGGCACGCCCCTCCTCATCACCGGG  135568737
-
-Query  61         -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-                   |||||||||||||||||||||||||||||||||||||| ||||||||||||||||||||
-Sbjct  135568738  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGAGGCAGGCCAGCTGAGTCCTG  135568797
-
-Query  120        AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-                  ||||| |||||| | ||||||| ||||| ||||||||||||||||||||||||||||  |
-Sbjct  135568798  AGCAGGAGCCCAACACAGCCACTGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCT  135568857
-
-Query  180        GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTG--CCCCCACCTCCCC-TC  236
-                  |||||||||| ||||||||||| ||||||||||||||||||   ||||||||||||| ||
-Sbjct  135568858  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTGAGTCGCCCCCCACCTCCCCCTC  135568917
-
-Query  237        AGGCCGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCC  296
-                  ||||||||  ||||||||||||||||  ||||||||||||||||||||||||||||||||
-Sbjct  135568918  AGGCCGCACAGCAGTGGGGGCTGAGAAAAGGAAGCACCATGGCCCACCTCTTCTCACCCC  135568977
-
-Query  297        TTTGGCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCC  356
-                  |||||||||||||||||||||||||||| |||| ||||||||||||||||||||||||||
-Sbjct  135568978  TTTGGCTGGCAGTCCCTTTGCAGTCTAATCACCCTGTTGCAGGCTCAATCCATTTGCCCC  135569037
-
-Query  357        AGCTCTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGG  416
-                  |||||||||||||||||||||||||| |||||||||||||||||||||| ||||||||||
-Sbjct  135569038  AGCTCTGCCCTTGCAGAGGGAGAGGAAGGAAGAGCAAGCTGCCCGAGACACAGGGGAAGG  135569097
-
-Query  417        AGGATGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCG  476
-                  |||||||| ||||||||||||||| |||||||||||||||||||||||||||||||||||
-Sbjct  135569098  AGGATGAGAGCCCTGGGGATGAGCAGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCG  135569157
-
-Query  477        AAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCC  536
-                  || |||||||||||||||||||||||||||||||||| |||||||||||||||||| |||
-Sbjct  135569158  AAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTAAGAGGATGGACCTGATGGGCTCC  135569217
-
-Query  537        TGGACCCTCCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTG  596
-                  ||||||||||| ||||||| |||||||||||||||||||||| |||||||||||||||||
-Sbjct  135569218  TGGACCCTCCCTTCTCACCTTGGTCCCTCAGTCTCATTCCCCTACTCCTGCCACCTCCTG  135569277
-
-Query  597        TCTGGCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCT  656
-                  ||||||  ||||||||||||||||||| || |||||||||||||||||||||||||||||
-Sbjct  135569278  TCTGGCTGTCAGGAAGGCCAGCCTGCTGCCAACCTGATCCTCCCAAACCCAGAGCCACCT  135569337
-
-Query  657        GATGCCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGA  716
-                  ||| ||||||||||||||||||||||||||||||||||||||||| ||||| ||| ||||
-Sbjct  135569338  GATTCCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGNCAGCGCGGTGGTGA  135569397
-
-Query  717        AGAGACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCA  776
-                  ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  135569398  AGAGACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCA  135569457
-
-Query  777        AGGCCCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGG  836
-                  ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  135569458  AGGCCCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGG  135569517
-
-Query  837        TTGTGGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGA  896
-                   ||||||||||| |||||||||| ||||||||||| ||||||||||||||||||||| ||
-Sbjct  135569518  CTGTGGTGGGGGAACAGGCAGCCAGCCCTGGTGGGGACCCTGGAGCCCCATGTGTAGTGA  135569577
-
-Query  897        GAGGAGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCC  956
-                  ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  135569578  GAGGAGGGATGGGCATTTTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCC  135569637
-
-Query  957        CAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACT  1016
-                  ||| ||||||||| |||||||||||||||| |||||||||||||||||||||||||||||
-Sbjct  135569638  CAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATCCGGACT  135569697
-
-Query  1017       GTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGG  1076
-                  |||||||| ||||||||||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  135569698  GTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACGGCCCGG  135569757
-
-Query  1077       TCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACC  1136
-                  ||||||||||||||| ||||||||||||||||| ||||||||||||||||||||||| ||
-Sbjct  135569758  TCTAGGGTGTCGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCAGGCCCC  135569817
-
-Query  1137       CTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCC  1196
-                  ||||||||| ||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  135569818  CTTCTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGGGGTCCC  135569877
-
-Query  1197       CATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-                  ||||     ||||||||||||||||||| || |||
-Sbjct  135569878  CATC-NNNNAGCTGCTCCCAAATAAACTTCAAAAG  135569911
-
-
->DQ977353.1 Pan troglodytes BGLAP (BGLAP) gene, complete cds
-Length=1064
-
- Score = 1896 bits (2102),  Expect = 0.0
- Identities = 1059/1064 (99%), Gaps = 0/1064 (0%)
- Strand=Plus/Plus
-
-Query  74    GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  133
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1     GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  60
-
-Query  134   GCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCAC  193
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  61    GCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCAC  120
-
-Query  194   TTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  253
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  121   TTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  180
-
-Query  254   GGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  313
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  181   GGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  240
-
-Query  314   TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  373
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  241   TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  300
-
-Query  374   GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGG  433
-             |||||||||||||||||||||||||||||||| |||||||||||||||||||||||||||
-Sbjct  301   GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACACAGGGGAAGGAGGATGAGGGCCCTGGG  360
-
-Query  434   GATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGA  493
-             ||||||||||||||| ||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  361   GATGAGCTGGGGTGAGCCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGA  420
-
-Query  494   GTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  553
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  421   GTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  480
-
-Query  554   CCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGG  613
-             ||||| ||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  481   CCCTGTTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGG  540
-
-Query  614   CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  673
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  541   CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  600
-
-Query  674   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTAC  733
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  601   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTAC  660
-
-Query  734   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  793
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  661   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  720
-
-Query  794   GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAG  853
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  721   GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAG  780
-
-Query  854   GCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATT  913
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  781   GCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATT  840
-
-Query  914   TTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGAT  973
-             |||||| |||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  841   TTGCACAGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGAT  900
-
-Query  974   CCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGAC  1033
-              |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  901   ACCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGAC  960
-
-Query  1034  CACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTG  1093
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  961   CACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTG  1020
-
-Query  1094  CTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCC  1137
-             ||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1021  CTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCC  1064
-
-
->DQ976476.1 Gorilla gorilla BGLAP (BGLAP) gene, complete cds
-Length=1069
-
- Score = 1878 bits (2082),  Expect = 0.0
- Identities = 1061/1073 (99%), Gaps = 4/1073 (0%)
- Strand=Plus/Plus
-
-Query  75    TATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCG  134
-             |||||||| | |||||||||||| |||||||||||||||||||||||||||||||||| |
-Sbjct  1     TATAAACANTNCTGGAGGCTGGCAGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGTG  60
-
-Query  135   CAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACT  194
-             |||||||||||||||||||||||| |||||||||||||||||||||||||||||||||||
-Sbjct  61    CAGCCACCGAGACACCATGAGAGCTCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACT  120
-
-Query  195   TTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGGG  254
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  121   TTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGGG  180
-
-Query  255   GGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCTT  314
-             |||||||||||||||||||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  181   GGCTGAGAGGAGGAAGCACCATGGCCCATCTCTTCTCACCCCTTTGGCTGGCAGTCCCTT  240
-
-Query  315   TGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGAG  374
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  241   TGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGAG  300
-
-Query  375   GGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGGG  434
-             |||||||||||||||||||||||||||||||||||||||||||   ||||||||||||||
-Sbjct  301   GGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGA---TGAGGGCCCTGGGG  357
-
-Query  435   ATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGAG  494
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  358   ATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGAG  417
-
-Query  495   TCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCAC  554
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  418   TCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCAC  477
-
-Query  555   CCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGGC  614
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  478   CCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGGC  537
-
-Query  615   CAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGCT  674
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  538   CAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGCT  597
-
-Query  675   CCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACC  734
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  598   CCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACC  657
-
-Query  735   TGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGGG  794
-             |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  658   TGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGGTGGGCCAAGGCCCTGCCTCTCCGGG  717
-
-Query  795   ATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAGG  854
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  718   ATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAGG  777
-
-Query  855   CAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATTT  914
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  778   CAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATTT  837
-
-Query  915   TGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGATC  974
-             |||||||||||||||||||||| |||||||||||||||||||||||||||||||||||||
-Sbjct  838   TGCACGGGGGCTGATGCCACCATGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGATC  897
-
-Query  975   CCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGACC  1034
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  898   CCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGACC  957
-
-Query  1035  ACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTGC  1094
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  958   ACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTGC  1017
-
-Query  1095  TGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCCTTCTTTCCTC  1147
-             |||||||||||||||||||||||||||||||||||||||| ||||||||||||
-Sbjct  1018  TGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCA-CCTTCTTTCCTC  1069
-
-
->DQ977503.1 Pongo pygmaeus BGLAP (BGLAP) gene, complete cds
-Length=1064
-
- Score = 1838 bits (2038),  Expect = 0.0
- Identities = 1046/1064 (98%), Gaps = 0/1064 (0%)
- Strand=Plus/Plus
-
-Query  74    GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  133
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1     GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  60
-
-Query  134   GCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCAC  193
-             |||||||||||||| ||||||||||||||||||||||||||||| |||||||||||||||
-Sbjct  61    GCAGCCACCGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTACTGGCCCTGGCCGCAC  120
-
-Query  194   TTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  253
-             | |||||| |||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  121   TGTGCATCACTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  180
-
-Query  254   GGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  313
-             ||||||||||||||||||||||  ||||||||||||||||||||||||||||||||||||
-Sbjct  181   GGGCTGAGAGGAGGAAGCACCACAGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  240
-
-Query  314   TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  373
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  241   TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  300
-
-Query  374   GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGG  433
-             ||||||||||||||||||||||||||||||| |||||| |||||||||||||||||||||
-Sbjct  301   GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGATGCAGGGAAAGGAGGATGAGGGCCCTGGG  360
-
-Query  434   GATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGA  493
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  361   GATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGA  420
-
-Query  494   GTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  553
-              |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  421   CTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  480
-
-Query  554   CCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGG  613
-             |||||||||||||||||||||||||||||||||||||||||||||||||||||||||| |
-Sbjct  481   CCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAAG  540
-
-Query  614   CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  673
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  541   CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  600
-
-Query  674   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTAC  733
-             |||||||||||||||||||||||||||||||||||||| |||||||||||||||||||||
-Sbjct  601   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTAC  660
-
-Query  734   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  793
-             ||||||||||||||||||||||||||||||| ||||||||||||||||||||||||| ||
-Sbjct  661   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCGGAGCTGGGCCAAGGCCCTGCCTCTCTGG  720
-
-Query  794   GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAG  853
-             ||||||||||||||||| ||||||||||||||||||||||||||||||||||||||||| 
-Sbjct  721   GATGGTCTGTGGGGGAGGTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAC  780
-
-Query  854   GCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATT  913
-             |||||||||||||||||||||||||||||||||||||||| |||||||||||||||||||
-Sbjct  781   GCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGTGAGAGGAGGGATGGGCATT  840
-
-Query  914   TTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGAT  973
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  841   TTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGAT  900
-
-Query  974   CCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGAC  1033
-             ||||||||||||| ||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  901   CCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGAC  960
-
-Query  1034  CACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTG  1093
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  961   CACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTG  1020
-
-Query  1094  CTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCC  1137
-             ||||||||| ||||||||||||||||||||||||||||||||||
-Sbjct  1021  CTGGCCTGGTCGGCAACCCCAGTTCTGCTCCTCTCCAGGCACCC  1064
-
-
->DQ977098.1 Macaca nemestrina BGLAP (BGLAP) gene, complete cds
-Length=1059
-
- Score = 1712 bits (1898),  Expect = 0.0
- Identities = 1015/1059 (96%), Gaps = 0/1059 (0%)
- Strand=Plus/Plus
-
-Query  74    GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  133
-             ||||||||||||||||||||||||||||||||||||||||||||||||||| |||||| |
-Sbjct  1     GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGGAGCCCAAC  60
-
-Query  134   GCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCAC  193
-              ||||||| ||||| ||||||||||||||||||||||||||||  ||||||||||| |||
-Sbjct  61    ACAGCCACTGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCTGGCCCTGGCCACAC  120
-
-Query  194   TTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  253
-             |||||||| ||||||||||||||||||||||||||||||||||||||||||  |||||||
-Sbjct  121   TTTGCATCACTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCACAGCAGTGG  180
-
-Query  254   GGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  313
-             |||||||||  |||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  181   GGGCTGAGAAAAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  240
-
-Query  314   TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  373
-             ||||||||||| || | |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  241   TTGCAGTCTAATCATCCTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  300
-
-Query  374   GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGG  433
-             ||||||||| |||||||||||||||||||||| |||||||||||||||||| ||||||||
-Sbjct  301   GGGAGAGGAAGGAAGAGCAAGCTGCCCGAGACACAGGGGAAGGAGGATGAGAGCCCTGGG  360
-
-Query  434   GATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGA  493
-             ||||||| ||||||||||||||||||||||||||||||||||||| ||||||||||||||
-Sbjct  361   GATGAGCAGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAACCCAGCGGTGCAGA  420
-
-Query  494   GTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  553
-             |||||||||||||||||||| |||||||||||||||||| |||||||||||||| |||||
-Sbjct  421   GTCCAGCAAAGGTGCAGGTAAGAGGATGGACCTGATGGGCTCCTGGACCCTCCCTTCTCA  480
-
-Query  554   CCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGG  613
-             || |||||||||||||||||||||| |||||||||||||||||||||||  |||||||||
-Sbjct  481   CCTTGGTCCCTCAGTCTCATTCCCCTACTCCTGCCACCTCCTGTCTGGCTGTCAGGAAGG  540
-
-Query  614   CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  673
-             |||||||||| || |||||||||||||||||||||||||||||||| |||||||||||||
-Sbjct  541   CCAGCCTGCTGCCAACCTGATCCTCCCAAACCCAGAGCCACCTGATTCCTGCCCCTCTGC  600
-
-Query  674   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTAC  733
-             |||||||||||||||||||||||||||||||||||||| |||||||||||||||||||||
-Sbjct  601   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTAC  660
-
-Query  734   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  793
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  661   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  720
-
-Query  794   GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAG  853
-             ||||||||||||||||||||||||||||||||||||||||||| ||||||||||| ||||
-Sbjct  721   GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGCTGTGGTGGGGGAACAG  780
-
-Query  854   GCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATT  913
-             |||||| ||||||||||| ||||||||||||||||||||| |||||||||||||||||||
-Sbjct  781   GCAGCCAGCCCTGGTGGGGACCCTGGAGCCCCATGTGTAGTGAGAGGAGGGATGGGCATT  840
-
-Query  914   TTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGAT  973
-             |||||||||||||||||||||||||||||||||||||||||||||| ||||||||| |||
-Sbjct  841   TTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGCCCCCTACCCAGAT  900
-
-Query  974   CCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGAC  1033
-             ||||||||||||| ||||||||||||||||||||||||||||||||||||| ||||||||
-Sbjct  901   CCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGCTGGCTGAC  960
-
-Query  1034  CACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTG  1093
-             ||||||||||| |||||||||||||||||||||||||||||||||||||||||||||| |
-Sbjct  961   CACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCCG  1020
-
-Query  1094  CTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCAGG  1132
-             |||||||||||||||| ||||||||||||||||||||||
-Sbjct  1021  CTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCAGG  1059
-
-
->DQ976615.1 Ateles geoffroyi BGLAP (BGLAP) gene, complete cds
-Length=1026
-
- Score = 1642 bits (1820),  Expect = 0.0
- Identities = 979/1026 (95%), Gaps = 0/1026 (0%)
- Strand=Plus/Plus
-
-Query  74    GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  133
-             ||||||||| ||||||||||||||||||||||||||||||||||||||||| |||||| |
-Sbjct  1     GTATAAACANTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGGAGCCCAAC  60
-
-Query  134   GCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCAC  193
-              ||||||| ||||| ||||||||||||||||||||||||||||  ||||||||||| |||
-Sbjct  61    ACAGCCACTGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCTGGCCCTGGCCACAC  120
-
-Query  194   TTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  253
-             |||||||| ||||||||||||||||||||||||||||||||||||||||||  |||||||
-Sbjct  121   TTTGCATCACTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCACAGCAGTGG  180
-
-Query  254   GGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  313
-             |||||||||  |||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  181   GGGCTGAGAAAAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  240
-
-Query  314   TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  373
-             ||||||||||| || | |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  241   TTGCAGTCTAATCATCCTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  300
-
-Query  374   GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGG  433
-             ||||||||| |||||||||||||||||||||| |||||||||||||||||| ||||||||
-Sbjct  301   GGGAGAGGAAGGAAGAGCAAGCTGCCCGAGACACAGGGGAAGGAGGATGAGAGCCCTGGG  360
-
-Query  434   GATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGA  493
-             ||||||| ||||||||||||||||||||||||||||||||||||| ||||||||||||||
-Sbjct  361   GATGAGCAGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAACCCAGCGGTGCAGA  420
-
-Query  494   GTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  553
-             |||||||||||||||||||| |||||||||||||||||| |||||||||||||| |||||
-Sbjct  421   GTCCAGCAAAGGTGCAGGTAAGAGGATGGACCTGATGGGCTCCTGGACCCTCCCTTCTCA  480
-
-Query  554   CCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGG  613
-             || |||||||||||||||||||||| |||||||||||||||||||||||  |||||||||
-Sbjct  481   CCTTGGTCCCTCAGTCTCATTCCCCTACTCCTGCCACCTCCTGTCTGGCTGTCAGGAAGG  540
-
-Query  614   CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  673
-             |||||||||| || |||||||||||||||||||||||||||||||| |||||||||||||
-Sbjct  541   CCAGCCTGCTGCCAACCTGATCCTCCCAAACCCAGAGCCACCTGATTCCTGCCCCTCTGC  600
-
-Query  674   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTAC  733
-             |||||||||||||||||||||||||||||||||||||| |||||||||||||||||||||
-Sbjct  601   TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTAC  660
-
-Query  734   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  793
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  661   CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  720
-
-Query  794   GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAG  853
-             ||||||||||||||||||||||||||||||||||||||||||| ||||||||||| ||||
-Sbjct  721   GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGCTGTGGTGGGGGAACAG  780
-
-Query  854   GCAGCCTGCCCTGGTGGGCACCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATT  913
-             ||||||    |||||||| ||||||||||||||||||||| |||||||||||||||||||
-Sbjct  781   GCAGCCANNNCTGGTGGGGACCCTGGAGCCCCATGTGTAGTGAGAGGAGGGATGGGCATT  840
-
-Query  914   TTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGAT  973
-             |||||||||||||||||||||||||||||||||||||||||||||| ||||||||| |||
-Sbjct  841   TTGCACGGGGGCTGATGCCACCACGTCGGGTGTCTCAGAGCCCCAGCCCCCTACCCAGAT  900
-
-Query  974   CCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGAC  1033
-             ||||||||||||| ||||||||||||||||||||||||||||||||||||| ||||||||
-Sbjct  901   CCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATCCGGACTGTGACGAGCTGGCTGAC  960
-
-Query  1034  CACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTG  1093
-             ||||||||||| |||||||||||||||||||||||||||||||||||||||||||||| |
-Sbjct  961   CACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCCG  1020
-
-Query  1094  CTGGCC  1099
-             ||||||
-Sbjct  1021  CTGGCC  1026
-
-
->DQ976876.1 Macaca mulatta BGLAP (BGLAP) gene, complete cds
-Length=1010
-
- Score = 1626 bits (1802),  Expect = 0.0
- Identities = 969/1014 (96%), Gaps = 4/1014 (0%)
- Strand=Plus/Plus
-
-Query  94    TGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGCCACCGAGACACCATG  153
-             ||||||||||||||||||||||||||||||| |||||| | ||||||| ||||| |||||
-Sbjct  1     TGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGGAGCCCAACACAGCCACTGAGACGCCATG  60
-
-Query  154   AGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCA  213
-             |||||||||||||||||||||||  ||||||||||| ||||||||||| |||||||||||
-Sbjct  61    AGAGCCCTCACACTCCTCGCCCTGCTGGCCCTGGCCACACTTTGCATCACTGGCCAGGCA  120
-
-Query  214   GGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGGGGGCTGAGAGGAGGAAGCAC  273
-             |||||||||||||||||||||||||||||||  ||||||||||||||||  |||||||||
-Sbjct  121   GGTGAGTGCCCCCACCTCCCCTCAGGCCGCACAGCAGTGGGGGCTGAGAAAAGGAAGCAC  180
-
-Query  274   CATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCTTTGCAGTCTAACCACCTTGT  333
-             ||||||||||||||||||||||||||||||||||||||||||||||||||| |||| |||
-Sbjct  181   CATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCTTTGCAGTCTAATCACCCTGT  240
-
-Query  334   TGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAA  393
-             ||||||||||||||||||||||||||||||||||||||||||||||||| ||||||||||
-Sbjct  241   TGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGAGGGAGAGGAAGGAAGAGCAA  300
-
-Query  394   GCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGGGATGAGCTGGGGTGAACCAG  453
-             |||||||||||| |||||||||||||||||| ||||||||||||||| ||||||||||||
-Sbjct  301   GCTGCCCGAGACACAGGGGAAGGAGGATGAGAGCCCTGGGGATGAGCAGGGGTGAACCAG  360
-
-Query  454   GCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTA  513
-             ||||||||||||||||||||||||| ||||||||||||||||||||||||||||||||||
-Sbjct  361   GCTCCCTTTCCTTTGCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTA  420
-
-Query  514   TGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCACCCTGGTCCCTCAGTCTCAT  573
-              |||||||||||||||||| |||||||||||    ||||||| |||||||||||||||||
-Sbjct  421   AGAGGATGGACCTGATGGGCTCCTGGACCCT----TCTCACCTTGGTCCCTCAGTCTCAT  476
-
-Query  574   TCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGGCCAGCCTGCTCCCCACCTGA  633
-             ||||| |||||||||||||||||||||||  ||||||||||||||||||| || ||||||
-Sbjct  477   TCCCCTACTCCTGCCACCTCCTGTCTGGCTGTCAGGAAGGCCAGCCTGCTGCCAACCTGA  536
-
-Query  634   TCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAA  693
-             |||||||||||||||||||||||||| |||||||||||||||||||||||||||||||||
-Sbjct  537   TCCTCCCAAACCCAGAGCCACCTGATTCCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAA  596
-
-Query  694   GCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGTATCAATGGCTGGGGTG  753
-             |||||||||||||||||| |||||||||||||||||||||||||||||||||||||||||
-Sbjct  597   GCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGTATCAATGGCTGGGGTG  656
-
-Query  754   AGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTG  813
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  657   AGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTG  716
-
-Query  814   CAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCA  873
-             ||||||||||||||||||||||| ||||||||||| |||||||||| ||||||||||| |
-Sbjct  717   CAGCAGGGAGTGGCCTCTCTGGGCTGTGGTGGGGGAACAGGCAGCCAGCCCTGGTGGGGA  776
-
-Query  874   CCCTGGAGCCCCATGTGTAGGGAGAGGAGGGATGGGCATTTTGCACGGGGGCTGATGCCA  933
-             |||||||||||||||||||| |||||||||||||||||||||||||||||||||||||||
-Sbjct  777   CCCTGGAGCCCCATGTGTAGTGAGAGGAGGGATGGGCATTTTGCACGGGGGCTGATGCCA  836
-
-Query  934   CCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGA  993
-             |||||||||||||||||||||||||| ||||||||| |||||||||||||||| ||||||
-Sbjct  837   CCACGTCGGGTGTCTCAGAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGA  896
-
-Query  994   GGTGTGTGAGCTCAATCCGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGC  1053
-             ||||||||||||||||||||||||||||||| ||||||||||||||||||| ||||||||
-Sbjct  897   GGTGTGTGAGCTCAATCCGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGC  956
-
-Query  1054  CTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGC  1107
-             |||||||||||||||||||||||||||||||||||||| |||||||||||||||
-Sbjct  957   CTATCGGCGCTTCTACGGCCCGGTCTAGGGTGTCGCTCCGCTGGCCTGGCCGGC  1010
-
-
->DQ977209.1 Pan paniscus BGLAP (BGLAP) gene, partial cds
-Length=790
-
- Score = 1409 bits (1562),  Expect = 0.0
- Identities = 786/790 (99%), Gaps = 0/790 (0%)
- Strand=Plus/Plus
-
-Query  74   GTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  133
-            ||||||||| | ||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1    GTATAAACANTNCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGC  60
-
-Query  134  GCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCAC  193
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  61   GCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCAC  120
-
-Query  194  TTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  253
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  121  TTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTGG  180
-
-Query  254  GGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  313
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  181  GGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCT  240
-
-Query  314  TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  373
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  241  TTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAGA  300
-
-Query  374  GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGG  433
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  301  GGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGG  360
-
-Query  434  GATGAGCTGGGGTGAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGA  493
-            ||||||||||||||| ||||||||||||||||||||||||||||| ||||||||||||||
-Sbjct  361  GATGAGCTGGGGTGAGCCAGGCTCCCTTTCCTTTGCAGGTGCGAAACCCAGCGGTGCAGA  420
-
-Query  494  GTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  553
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  421  GTCCAGCAAAGGTGCAGGTATGAGGATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCA  480
-
-Query  554  CCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGG  613
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  481  CCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGG  540
-
-Query  614  CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  673
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  541  CCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGC  600
-
-Query  674  TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTAC  733
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  601  TCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTAC  660
-
-Query  734  CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  793
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  661  CTGTATCAATGGCTGGGGTGAGAGAAAAGGCAGAGCTGGGCCAAGGCCCTGCCTCTCCGG  720
-
-Query  794  GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAG  853
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  721  GATGGTCTGTGGGGGAGCTGCAGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAG  780
-
-Query  854  GCAGCCTGCC  863
-            ||||||||||
-Sbjct  781  GCAGCCTGCC  790
-
-
->DQ976754.1 Lagothrix lagotricha BGLAP (BGLAP) gene, partial cds
-Length=634
-
- Score = 758 bits (840),  Expect = 0.0
- Identities = 574/677 (85%), Gaps = 52/677 (8%)
- Strand=Plus/Plus
-
-Query  74   GTATAAACAGTGCTGGAGGCTGGCGGGG--CAGGCCAGCTGAGTCCTGAGCAGCAGCCCA  131
-            ||||||||||||||||||| || |||||  ||||| ||| ||| | |||||||||||   
-Sbjct  1    GTATAAACAGTGCTGGAGGGTGCCGGGGGGCAGGCGAGCCGAGCCTTGAGCAGCAGCTG-  59
-
-Query  132  GCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGC  191
-                     ||||||| |||||||||||||| |||||||| ||||  |||||||||||||
-Sbjct  60   ---------CCGAGACGCCATGAGAGCCCTCCCACTCCTCCCCCTGCTGGCCCTGGCCGC  110
-
-Query  192  ACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGT  251
-            ||| |||||||||||||||||||||||||| |||||||||||||||||||||| ||||||
-Sbjct  111  ACTCTGCATCGCTGGCCAGGCAGGTGAGTGTCCCCACCTCCCCTCAGGCCGCACTGCAGT  170
-
-Query  252  GGGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCC  311
-            ||||||||||||| | |   ||||||||||||||||||||||||||| | ||||||||||
-Sbjct  171  GGGGGCTGAGAGGGGTAT--ACCATGGCCCACCTCTTCTCACCCCTTCG-CTGGCAGTCC  227
-
-Query  312  CTTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCA  371
-            ||                  ||||||||||||| ||||||| ||  ||||||||||||||
-Sbjct  228  CT------------------GTTGCAGGCTCAAGCCATTTGTCCTCGCTCTGCCCTTGCA  269
-
-Query  372  GAGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTG  431
-            ||| |||||||||||||| || ||||||| ||| ||||||||||||||||||||||||| 
-Sbjct  270  GAGAGAGAGGAGGGAAGAACACGCTGCCCAAGATGCAGGGGAAGGAGGATGAGGGCCCTA  329
-
-Query  432  GGGATGAGCTGGGGT-----GAACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCG  486
-            ||| ||||| ||| |     |||| | |||||||| ||||||||||||||||||||||||
-Sbjct  330  GGGGTGAGC-GGGTTCAAAGGAACTATGCTCCCTT-CCTTTGCAGGTGCGAAGCCCAGCG  387
-
-Query  487  GTGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTG--ATGGGTTCCTGGACCCT  544
-            | |||||||||||||||||||||||||||||||||| || |  | ||| |||||||||||
-Sbjct  388  GCGCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGGCCCGGCAGGGGCTCCTGGACCCT  447
-
-Query  545  CCCCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCA  604
-            ||||||||||||| |||||||||||||||||||| ||||||||||          |||  
-Sbjct  448  CCCCTCTCACCCTAGTCCCTCAGTCTCATTCCCCAACTCCTGCCA----------GGCTG  497
-
-Query  605  TCAGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTG  664
-            |||||||||||| ||||||||| ||||  |||||||||||||||||||||||||||| | 
-Sbjct  498  TCAGGAAGGCCACCCTGCTCCCTACCTTGTCCTCCCAAACCCAGAGCCACCTGATGCTTA  557
-
-Query  665  CCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCC  724
-            ||||||||||||||||||||||||||||||||||||||||||||||| ||||||||||||
-Sbjct  558  CCCCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCC  617
-
-Query  725  AGGCGCTACCTGTATCA  741
-            ||||| |||||||||||
-Sbjct  618  AGGCGTTACCTGTATCA  634
-
-
->DQ976983.1 Saguinus labiatus BGLAP (BGLAP) gene, partial cds
-Length=630
-
- Score = 724 bits (802),  Expect = 0.0
- Identities = 563/672 (84%), Gaps = 49/672 (7%)
- Strand=Plus/Plus
-
-Query  74   GTATAAACAGTGCTGGAGGCTGGC-GGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAG  132
-            ||||||||||||| ||||| || | ||||||||||||| ||| |||| ||||||||    
-Sbjct  1    GTATAAACAGTGCCGGAGGGTGCCAGGGGCAGGCCAGCCGAGCCCTGCGCAGCAGCTG--  58
-
-Query  133  CGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCA  192
-                    ||||||| ||||||||||||||||||||||| |  |  ||||||||||||||
-Sbjct  59   --------CCGAGACGCCATGAGAGCCCTCACACTCCTCCCTGTGCTGGCCCTGGCCGCA  110
-
-Query  193  CTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGCCGCATTGCAGTG  252
-            ||||||||||||||||||||||||||||| ||||||||||||||||||| || |||||||
-Sbjct  111  CTTTGCATCGCTGGCCAGGCAGGTGAGTGTCCCCACCTCCCCTCAGGCCCCACTGCAGTG  170
-
-Query  253  GGGGCTGAGAGGAGGAAGCACCATGGCCCACCTCTTCTCACCCCTTTGGCTGGCAGTCCC  312
-            |||||||||||| | |   |||| ||| |||||||||| ||||||| | |||||||||||
-Sbjct  171  GGGGCTGAGAGGGGTAT--ACCACGGCTCACCTCTTCTTACCCCTTGGCCTGGCAGTCCC  228
-
-Query  313  TTTGCAGTCTAACCACCTTGTTGCAGGCTCAATCCATTTGCCCCAGCTCTGCCCTTGCAG  372
-            |                  |||| |||||||| ||||||| || |||| |||||||||||
-Sbjct  229  T------------------GTTGTAGGCTCAAGCCATTTGTCCTAGCTGTGCCCTTGCAG  270
-
-Query  373  AGGGAGAGGAGGGAAGAGCAAGCTGCCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGG  432
-            ||||||||||||||||| ||  ||||| ||| |  ||||||||||||||||||||||| |
-Sbjct  271  AGGGAGAGGAGGGAAGAACACACTGCCTGAGCCATAGGGGAAGGAGGATGAGGGCCCTCG  330
-
-Query  433  GGATGAGCTGGGGTGA----ACCAGGCTCCCTTTCCTTTGCAGGTGCGAAGCCCAGCGGT  488
-            || ||||| ||| | |    |||| |||||||| ||||| ||||||||||||||||| ||
-Sbjct  331  GGGTGAGCAGGGTTCAGAGGACCATGCTCCCTT-CCTTTTCAGGTGCGAAGCCCAGCAGT  389
-
-Query  489  GCAGAGTCCAGCAAAGGTGCAGGTATGAGGATGGACCTG--ATGGGTTCCTGGACCCTCC  546
-            |||||||||||||||||| ||||||||||||||| || |  | ||| |||||||||||||
-Sbjct  390  GCAGAGTCCAGCAAAGGTACAGGTATGAGGATGGGCCCGGCAGGGGCTCCTGGACCCTCC  449
-
-Query  547  CCTCTCACCCTGGTCCCTCAGTCTCATTCCCCCACTCCTGCCACCTCCTGTCTGGCCATC  606
-            |||||||||||||||||||||||| ||||||||||||||||||          |||  ||
-Sbjct  450  CCTCTCACCCTGGTCCCTCAGTCTTATTCCCCCACTCCTGCCA----------GGCTGTC  499
-
-Query  607  AGGAAGGCCAGCCTGCTCCCCACCTGATCCTCCCAAACCCAGAGCCACCTGATGCCTGCC  666
-            |||||||||| ||||||||| ||||  ||| |||||||||||||||||||||||| ||||
-Sbjct  500  AGGAAGGCCACCCTGCTCCCTACCTTGTCC-CCCAAACCCAGAGCCACCTGATGCTTGCC  558
-
-Query  667  CCTCTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAG  726
-            ||| ||||||||||||||||||||||||||||||||||||||||| ||||||||||||||
-Sbjct  559  CCTTTGCTCCACAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAG  618
-
-Query  727  GCGCTACCTGTA  738
-            ||||||||||||
-Sbjct  619  GCGCTACCTGTA  630
-
-
->XM_020169839.1 PREDICTED: Castor canadensis polyamine-modulated factor 1-like 
-(LOC109690472), mRNA
-Length=3629
-
- Score = 517 bits (572),  Expect = 4e-142
- Identities = 715/979 (73%), Gaps = 74/979 (8%)
- Strand=Plus/Plus
-
-Query  44    CCCTCCT-CATCGCTGGGCACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGG-  101
-             ||||||| || | | ||| ||||||||| | |||||||||||| |||  |||||  ||| 
-Sbjct  2710  CCCTCCTGCACCTCAGGG-ACAGCCCAGCGTGTATAAACAGTGGTGGCTGCTGGAAGGGA  2768
-
-Query  102   CAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCT  161
-             ||| ||||||||||||||| ||||||||| | |||||   | |||||||||||| |||||
-Sbjct  2769  CAGCCCAGCTGAGTCCTGAACAGCAGCCCTGTGCAGCTGAC-AGACACCATGAGGGCCCT  2827
-
-Query  162   CACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTG  221
-             ||  || || |||||| ||||||||||| |||| ||| || |||||  | ||||||||| 
-Sbjct  2828  CATTCTGCTTGCCCTACTGGCCCTGGCCACACTCTGCCTCACTGGCTGGACAGGTGAGTA  2887
-
-Query  222   CCCCCACCTCCCCTCAGGCCGCATTGCAGTGGGGGCTG--AGAGGAGGAAGCACCATGGC  279
-              | |||   || |||||  |   |  || ||||| ||   ||||||| | ||  |||| |
-Sbjct  2888  TCACCAATCCCACTCAGATCTTCTC-CAATGGGG-CTACAAGAGGAGCATGC--CATGTC  2943
-
-Query  280   -CCACCTCTTCTCACCCCTTTGGCTGGCAGTCCCTTTGCAGTCTAACCACCTTGTTGCAG  338
-              ||| |||||     ||||||          || | |||| || | ||| |||  |||||
-Sbjct  2944  TCCAACTCTTA----CCCTTT----------CCTTCTGCAATCCAGCCATCTTCCTGCAG  2989
-
-Query  339   GCTCAATCCATTTG-CCCCAGCTCTGCCCTTGCAGAGGGAGAGGAGGGAAGAGCAAGCTG  397
-             ||||||||| |||| |||||| |||| |||||  |||||| || |||| |||||| ||||
-Sbjct  2990  GCTCAATCCTTTTGTCCCCAGTTCTGTCCTTGTGGAGGGACAG-AGGGCAGAGCATGCTG  3048
-
-Query  398   CCCGAGACGCAGGGGAAGGAGGATGAGGGCCCTGGGGATGAGCTGGGGTGAACCAGGCTC  457
-             ||  ||| ||||||| || ||| ||||| ||||  | | |  | |||||||||  || ||
-Sbjct  3049  CC--AGAAGCAGGGGCAGAAGGTTGAGG-CCCTAAGAAAGGACAGGGGTGAACT-GGTTC  3104
-
-Query  458   CCTTTCCTTTGCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAGGTATGAG  517
-             | |   |||||||| ||||||||||||||||||||||||| |||||||| |||||| |||
-Sbjct  3105  CTTCCTCTTTGCAGATGCGAAGCCCAGCGGTGCAGAGTCCGGCAAAGGTTCAGGTAGGAG  3164
-
-Query  518   GATGGACCTGATGGGTTCCTGGACCCTCCCCTCTCACCCTGGTCCC-TCAGTCTCATTCC  576
-             |  ||  ||||| |  |||  | |  | ||| |||||| || |||| ||||   ||| ||
-Sbjct  3165  GCCGGGTCTGATTGTCTCCCAGGC--TGCCC-CTCACCATGCTCCCCTCAGCAACAT-CC  3220
-
-Query  577   CCCACTCCTGCCACCTCCTGTCTGGCCATCAGGAAGGCCAGCCTGCTCCCCACCTGATCC  636
-             |||||||  ||  |||||||||||||  |||  || |||| || ||| |           
-Sbjct  3221  CCCACTCTGGCTCCCTCCTGTCTGGCTGTCAAAAAAGCCACCCAGCTTC-----------  3269
-
-Query  637   TCCCAAACCCAGAGCCACCTGATGCCTGCCCCTCTGCTCCACAGCCTTTGTGTCCAAGCA  696
-             ||||||||   |||||||||||| ||  |||||||||||||||||||||   ||||||||
-Sbjct  3270  TCCCAAACTGGGAGCCACCTGATCCCCACCCCTCTGCTCCACAGCCTTTACATCCAAGCA  3329
-
-Query  697   GGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGTATCAATGGCTGGGGTGAGA  756
-             ||||||||||||| | |||||||||| || |||||||||| | ||  | |||||||||||
-Sbjct  3330  GGAGGGCAGCGAGATGGTGAAGAGACTCAAGCGCTACCTGGACCATGGCCTGGGGTGAGA  3389
-
-Query  757   GAAAAGGCAGAGCTGGGCCAAGG--CCCTGCCTCTCCGGGATGGTCTGTGGGGGAGCTGC  814
-               |||||  | ||||||||| ||  |||||  ||||| |||||  |||  |  ||| || 
-Sbjct  3390  AGAAAGGAGGGGCTGGGCCAGGGTCCCCTG--TCTCCAGGATGTGCTGGAGAAGAGTTGT  3447
-
-Query  815   AGCAGGGAGTGGCCTCTCTGGGTTGTGGTGGGGGTACAGGCAGCCTGCCCTGGTGGGCAC  874
-              |  ||| |        ||||     |||| |||     ||||  | |||       |||
-Sbjct  3448  GGTGGGGTG--------CTGG-----GGTGAGGGATGGAGCAGGATACCC-------CAC  3487
-
-Query  875   CCTGGAGCCCCATGTGTAG--GGAGAGGAGGGATGGGCATTTTGCACGGGGGCTGATGCC  932
-             ||   || ||| |||| ||  ||| |   ||| ||| || ||||   |||| ||||||||
-Sbjct  3488  CCCCAAGTCCCTTGTGCAGTAGGA-ATTGGGGGTGGACACTTTGTCTGGGGACTGATGCC  3546
-
-Query  933   ACCACGTCGGGTGTCTCAGAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGG  992
-             |    || ||   ||||||||| |||| ||||||||| ||||||||||||||||||||||
-Sbjct  3547  ATTGTGTGGGTCTTCTCAGAGCTCCAGCCCCCTACCCAGATCCCCTGGAGCCCAGGAGGG  3606
-
-Query  993   AGGTGTGTGAGCTCAATCC  1011
-             | |||||||||||||| ||
-Sbjct  3607  AAGTGTGTGAGCTCAACCC  3625
-
-
->NG_053096.1 Homo sapiens BGLAP promoter region (LOC110013312) on chromosome 
-1
-Length=1623
-
- Score = 515 bits (570),  Expect = 1e-141
- Identities = 285/285 (100%), Gaps = 0/285 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1339  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  1398
-
-Query  61    CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1399  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  1458
-
-Query  121   GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1459  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  1518
-
-Query  181   GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  240
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1519  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCCCACCTCCCCTCAGGC  1578
-
-Query  241   CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCT  285
-             |||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1579  CGCATTGCAGTGGGGGCTGAGAGGAGGAAGCACCATGGCCCACCT  1623
-
-
->NM_199173.5 Homo sapiens bone gamma-carboxyglutamate protein (BGLAP), mRNA
-Length=760
-
- Score = 508 bits (562),  Expect = 2e-139
- Identities = 281/281 (100%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  443   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  502
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  503   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  562
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  563   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  622
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  623   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  682
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  683   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  723
-
-
- Score = 392 bits (434),  Expect = 1e-104
- Identities = 217/217 (100%), Gaps = 0/217 (0%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  121  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  180
-
-Query  61   CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  181  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  240
-
-Query  121  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  241  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  300
-
-Query  181  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||||||||||||||||||||||||||||||
-Sbjct  301  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  337
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  371  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  430
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  431  ATCAATGGCTGGG  443
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  331  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  373
-
-
->NM_001199662.1 Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLAP), transcript 
-variant 2, mRNA
-Length=1001
-
- Score = 508 bits (562),  Expect = 2e-139
- Identities = 281/281 (100%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  701   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  760
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  761   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  820
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  821   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  880
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  881   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  940
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  941   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  981
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  629  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  688
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  689  ATCAATGGCTGGG  701
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  589  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  631
-
-
->NM_001199664.1 Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLAP), transcript 
-variant 4, mRNA
-Length=794
-
- Score = 508 bits (562),  Expect = 2e-139
- Identities = 281/281 (100%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  494   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  553
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  554   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  613
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  614   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  673
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  674   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  733
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  734   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  774
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  422  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  481
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  482  ATCAATGGCTGGG  494
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  382  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  424
-
-
->NM_001199663.1 Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLAP), transcript 
-variant 3, mRNA
-Length=805
-
- Score = 508 bits (562),  Expect = 2e-139
- Identities = 281/281 (100%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  505   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  564
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  565   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  624
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  625   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  684
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  685   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  744
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  745   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  785
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  433  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  492
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  493  ATCAATGGCTGGG  505
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 40/40 (100%), Gaps = 0/40 (0%)
- Strand=Plus/Plus
-
-Query  471  GGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            ||||||||||||||||||||||||||||||||||||||||
-Sbjct  396  GGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  435
-
-
->NM_001199661.1 Homo sapiens PMF1-BGLAP readthrough (PMF1-BGLAP), transcript 
-variant 1, mRNA
-Length=940
-
- Score = 508 bits (562),  Expect = 2e-139
- Identities = 281/281 (100%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  640   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  699
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  700   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  759
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  760   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  819
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  820   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  879
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  880   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  920
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  568  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  627
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  628  ATCAATGGCTGGG  640
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  528  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  570
-
-
->BC033656.2 Homo sapiens polyamine-modulated factor 1, mRNA (cDNA clone IMAGE:3909927)
-Length=820
-
- Score = 508 bits (562),  Expect = 2e-139
- Identities = 281/281 (100%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  495   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  554
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  555   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  614
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  615   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  674
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  675   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  734
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  735   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  775
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  423  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  482
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  483  ATCAATGGCTGGG  495
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 40/40 (100%), Gaps = 0/40 (0%)
- Strand=Plus/Plus
-
-Query  471  GGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            ||||||||||||||||||||||||||||||||||||||||
-Sbjct  386  GGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  425
-
-
->XM_016928883.1 PREDICTED: Pan troglodytes bone gamma-carboxyglutamate protein 
-(BGLAP), mRNA
-Length=1983
-
- Score = 502 bits (556),  Expect = 9e-138
- Identities = 280/281 (99%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||| ||||||||||||||||||||||||||||||||||||
-Sbjct  1694  GAGCCCCAGTCCCCTACCCGGATACCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1753
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1754  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1813
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1814  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1873
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1874  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1933
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  1934  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1974
-
-
- Score = 392 bits (434),  Expect = 1e-104
- Identities = 217/217 (100%), Gaps = 0/217 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1372  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  1431
-
-Query  61    CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1432  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  1491
-
-Query  121   GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1492  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  1551
-
-Query  181   GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||||||||||||||||||||||||||||||
-Sbjct  1552  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  1588
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1622  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  1681
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1682  ATCAATGGCTGGG  1694
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1582  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1624
-
-
->XM_004026955.2 PREDICTED: Gorilla gorilla gorilla bone gamma-carboxyglutamate 
-protein (BGLAP), mRNA
-Length=786
-
- Score = 495 bits (548),  Expect = 1e-135
- Identities = 279/281 (99%), Gaps = 1/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  487   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  546
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  547   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  606
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  607   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  666
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             |||| |||||||||||||||||||||||||||||||||||||||||||||| ||||||||
-Sbjct  667   GGCA-CCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTAGGGATGTGG  725
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  726   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  766
-
-
- Score = 378 bits (418),  Expect = 3e-100
- Identities = 214/217 (99%), Gaps = 0/217 (0%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  165  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  224
-
-Query  61   CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-            ||||||||||||||||||||||||||||||||||||| ||||||||||||||||||||||
-Sbjct  225  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCAGGGCAGGCCAGCTGAGTCCTGA  284
-
-Query  121  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-            |||||||||||| ||||||||||||||||||||||||| |||||||||||||||||||||
-Sbjct  285  GCAGCAGCCCAGTGCAGCCACCGAGACACCATGAGAGCTCTCACACTCCTCGCCCTATTG  344
-
-Query  181  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||||||||||||||||||||||||||||||
-Sbjct  345  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  381
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  415  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  474
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  475  ATCAATGGCTGGG  487
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  375  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  417
-
-
->XM_002810036.3 PREDICTED: Pongo abelii bone gamma-carboxyglutamate (gla) protein 
-(BGLAP), mRNA
-Length=545
-
- Score = 493 bits (546),  Expect = 5e-135
- Identities = 278/281 (99%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             |||||||||||||||||||||||||||||||||||| |||||||||||||||||||||||
-Sbjct  245   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  304
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  305   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  364
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             |||||||||||||||||||||||||||||||| |||||||||||||||||||||||||||
-Sbjct  365   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGTCGGCAACCCCAGTTCTGCTCCTCTCCA  424
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||| ||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  425   GGCACCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  484
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  485   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  525
-
-
- Score = 233 bits (258),  Expect = 7e-57
- Identities = 135/139 (97%), Gaps = 0/139 (0%)
- Strand=Plus/Plus
-
-Query  79   AACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGC  138
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1    AACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGC  60
-
-Query  139  CACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGC  198
-            ||||||||| ||||||||||||||||||||||||||||| |||||||||||||||| |||
-Sbjct  61   CACCGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTACTGGCCCTGGCCGCACTGTGC  120
-
-Query  199  ATCGCTGGCCAGGCAGGTG  217
-            ||| |||||||||||||||
-Sbjct  121  ATCACTGGCCAGGCAGGTG  139
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  173  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  232
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  233  ATCAATGGCTGGG  245
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||| ||||||||||||||||
-Sbjct  133  GCAGGTGCGAAGCCCAGCGGTGCAGACTCCAGCAAAGGTGCAG  175
-
-
->XM_008974558.2 PREDICTED: Pan paniscus bone gamma-carboxyglutamate (gla) protein 
-(BGLAP), mRNA
-Length=1934
-
- Score = 477 bits (528),  Expect = 4e-130
- Identities = 274/281 (98%), Gaps = 6/281 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||| ||||||||||||||||||||||||||||||||||||
-Sbjct  1655  GAGCCCCAGTCCCCTACCCGGATACCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1714
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1715  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1774
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1775  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1834
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||      ||||||||||||||||||||||||||||||||||
-Sbjct  1835  GGCACCCTTCTTTCCTCTTC------CCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1888
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  1889  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1929
-
-
- Score = 387 bits (428),  Expect = 5e-103
- Identities = 216/217 (99%), Gaps = 0/217 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1333  GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  1392
-
-Query  61    CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1393  CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  1452
-
-Query  121   GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-             ||||||||||||| ||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1453  GCAGCAGCCCAGCACAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  1512
-
-Query  181   GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||||||||||||||||||||||||||||||
-Sbjct  1513  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  1549
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1583  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  1642
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1643  ATCAATGGCTGGG  1655
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             ||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  1543  GCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1585
-
-
->X51699.1 Human mRNA for bone Gla protein
- X53698.1 Human mRNA for Gla protein (BGP, osteocalcin)
-Length=451
-
- Score = 471 bits (522),  Expect = 2e-128
- Identities = 261/261 (100%), Gaps = 0/261 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  191   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  250
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  251   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  310
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  311   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  370
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  371   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  430
-
-Query  1191  GGTCCCCATCATCCCAGCTGC  1211
-             |||||||||||||||||||||
-Sbjct  431   GGTCCCCATCATCCCAGCTGC  451
-
-
- Score = 154 bits (170),  Expect = 6e-33
- Identities = 85/85 (100%), Gaps = 0/85 (0%)
- Strand=Plus/Plus
-
-Query  133  CGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCA  192
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1    CGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCA  60
-
-Query  193  CTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||||||||||||||||||
-Sbjct  61   CTTTGCATCGCTGGCCAGGCAGGTG  85
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  119  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  178
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  179  ATCAATGGCTGGG  191
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  79   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  121
-
-
->XM_003259455.1 PREDICTED: Nomascus leucogenys bone gamma-carboxyglutamate (gla) 
-protein (BGLAP), mRNA
-Length=743
-
- Score = 466 bits (516),  Expect = 7e-127
- Identities = 272/281 (97%), Gaps = 0/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||| |||| |||||||||||||||||||||||
-Sbjct  443   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAACCCAAGAGGGAGGTGTGTGAGCTCAATC  502
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| |||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  503   CGGACTGTGACGAGCTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  562
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             |||| ||||||||||| |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  563   GCCCAGTCTAGGGTGTTGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  622
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| |||||||| || |||| |||||||||||||||||||||||||||||||||||||||
-Sbjct  623   GGCCCCCTTCTTCCCCCTTCTCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  682
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  683   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  723
-
-
- Score = 356 bits (394),  Expect = 9e-94
- Identities = 209/217 (96%), Gaps = 0/217 (0%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            |||||||||||| |||||||||||||||||||||||||| ||||||||||||||||||||
-Sbjct  121  GGCAGATTCCCCGTAGACCCGCCCGCACCATGGTCAGGCGTGCCCCTCCTCATCGCTGGG  180
-
-Query  61   CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGA  120
-            |||||||||||||||||||||||||||||| |||||| ||||||||||||||||||||||
-Sbjct  181  CACAGCCCAGAGGGTATAAACAGTGCTGGAAGCTGGCAGGGCAGGCCAGCTGAGTCCTGA  240
-
-Query  121  GCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTG  180
-            |||||||||||||||||||||| |||| ||||||||||||||||||||||||||||  ||
-Sbjct  241  GCAGCAGCCCAGCGCAGCCACCAAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCTG  300
-
-Query  181  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||||||||||||||||||||||||||||||
-Sbjct  301  GCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  337
-
-
- Score = 123 bits (136),  Expect = 9e-24
- Identities = 71/73 (97%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| ||||||||||||||||||||||| |
-Sbjct  371  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTAT  430
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  431  ATCAATGGCTGGG  443
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  331  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  373
-
-
->XM_005541474.2 PREDICTED: Macaca fascicularis bone gamma-carboxyglutamate (gla) 
-protein (BGLAP), mRNA
-Length=1925
-
- Score = 450 bits (498),  Expect = 5e-122
- Identities = 269/281 (96%), Gaps = 1/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  1639  GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  1698
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1699  CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1758
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||| ||||||||||||||||| ||||||||||||||||||||
-Sbjct  1759  GCCCGGTCTAGGGTGTCGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCA  1818
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| ||||||||||| ||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1819  GGCCCCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1878
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||| ||||||||||||||||||||| || |||
-Sbjct  1879  GGTCCCCATCAT-CCAGCTGCTCCCAAATAAACTTCAAAAG  1918
-
-
- Score = 322 bits (356),  Expect = 2e-83
- Identities = 203/218 (93%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||| |||||||||||||||| ||||||||||||| | |||
-Sbjct  1316  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGGCACGCCCCTCCTCATCACCGGG  1375
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              |||||||||||||||||||||||||||||||||||||| ||||||||||||||||||||
-Sbjct  1376  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGAGGCAGGCCAGCTGAGTCCTG  1435
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             ||||| |||||| | ||||||| ||||| ||||||||||||||||||||||||||||  |
-Sbjct  1436  AGCAGGAGCCCAACACAGCCACTGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCT  1495
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||| ||||||||||| |||||||||||||||
-Sbjct  1496  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  1533
-
-
- Score = 123 bits (136),  Expect = 9e-24
- Identities = 71/73 (97%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||||||||||||||||||||||||||||| ||| |||||||||||||||||||||||||
-Sbjct  1567  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGCGGTGGTGAAGAGACCCAGGCGCTACCTGT  1626
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1627  ATCAATGGCTGGG  1639
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             ||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  1527  GCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1569
-
-
->XM_011769795.1 PREDICTED: Macaca nemestrina bone gamma-carboxyglutamate (gla) 
-protein (BGLAP), mRNA
-Length=1903
-
- Score = 450 bits (498),  Expect = 5e-122
- Identities = 269/281 (96%), Gaps = 1/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  1617  GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  1676
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1677  CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1736
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||| ||||||||||||||||| ||||||||||||||||||||
-Sbjct  1737  GCCCGGTCTAGGGTGTCGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCA  1796
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| ||||||||||| ||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1797  GGCCCCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1856
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||| ||||||||||||||||||||| || |||
-Sbjct  1857  GGTCCCCATCAT-CCAGCTGCTCCCAAATAAACTTCAAAAG  1896
-
-
- Score = 327 bits (362),  Expect = 4e-85
- Identities = 204/218 (94%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||| |||||||||||||||| ||||||||||||| | |||
-Sbjct  1294  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGGCACGCCCCTCCTCATCACCGGG  1353
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1354  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  1413
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             ||||| |||||| | ||||||| ||||| ||||||||||||||||||||||||||||  |
-Sbjct  1414  AGCAGGAGCCCAACACAGCCACTGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCT  1473
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||| ||||||||||| |||||||||||||||
-Sbjct  1474  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  1511
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1545  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1604
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1605  ATCAATGGCTGGG  1617
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             ||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  1505  GCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1547
-
-
->XM_011992748.1 PREDICTED: Mandrillus leucophaeus bone gamma-carboxyglutamate 
-(gla) protein (BGLAP), mRNA
-Length=744
-
- Score = 450 bits (498),  Expect = 5e-122
- Identities = 269/281 (96%), Gaps = 1/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  444   GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  503
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  504   CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  563
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||| ||||||||||||||||| ||||||||||||||||||||
-Sbjct  564   GCCCGGTCTAGGGTGTCGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCA  623
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| ||||||||||| ||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  624   GGCCCCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  683
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||| ||||||||||||||||||||| || |||
-Sbjct  684   GGTCCCCATCAT-CCAGCTGCTCCCAAATAAACTTCAAAAG  723
-
-
- Score = 322 bits (356),  Expect = 2e-83
- Identities = 203/218 (93%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            ||||||||||||||||||||||| |||||||||||||||| ||||||||||||| | |||
-Sbjct  121  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGGCACGCCCCTCCTCATCACCGGG  180
-
-Query  61   -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-             |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  181  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  240
-
-Query  120  AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-            ||||| |||||| | ||||||| ||||| ||||||||||| ||||||||||||||||  |
-Sbjct  241  AGCAGGAGCCCAACACAGCCACTGAGACGCCATGAGAGCCTTCACACTCCTCGCCCTGCT  300
-
-Query  180  GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||| ||||||||||| |||||||||||||||
-Sbjct  301  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  338
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  372  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  431
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  432  ATCAATGGCTGGG  444
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            ||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  332  GCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  374
-
-
->XM_012078677.1 PREDICTED: Cercocebus atys bone gamma-carboxyglutamate (gla) 
-protein (BGLAP), mRNA
-Length=1961
-
- Score = 441 bits (488),  Expect = 3e-119
- Identities = 267/281 (95%), Gaps = 1/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  1677  GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  1736
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1737  CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1796
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||| ||||||||||||||||| ||||||||||||||||||||
-Sbjct  1797  GCCCGGTCTAGGGTGTCGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCA  1856
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| | ||| ||||| ||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1857  GGCCCGCTTTTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1916
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||| ||||||||||||||||||||| || |||
-Sbjct  1917  GGTCCCCATCAT-CCAGCTGCTCCCAAATAAACTTCAAAAG  1956
-
-
- Score = 322 bits (356),  Expect = 2e-83
- Identities = 203/218 (93%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||| |||||||||||||||| ||||||||||||| | |||
-Sbjct  1354  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGGCACGCCCCTCCTCATCACCGGG  1413
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1414  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  1473
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             ||||| |||||| | ||||||| ||||| ||||||||||| ||||||||||||||||  |
-Sbjct  1474  AGCAGGAGCCCAACACAGCCACTGAGACGCCATGAGAGCCTTCACACTCCTCGCCCTGCT  1533
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||| ||||||||||| |||||||||||||||
-Sbjct  1534  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  1571
-
-
- Score = 123 bits (136),  Expect = 9e-24
- Identities = 71/73 (97%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||| ||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1605  CAGCTTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1664
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1665  ATCAATGGCTGGG  1677
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             ||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  1565  GCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1607
-
-
->XM_010366604.1 PREDICTED: Rhinopithecus roxellana osteocalcin (LOC104664919), 
-mRNA
-Length=545
-
- Score = 441 bits (488),  Expect = 3e-119
- Identities = 267/281 (95%), Gaps = 1/281 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  245   GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  304
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| ||||||||||||| |||||||||||
-Sbjct  305   CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCAGCGCTTCTACG  364
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||| |||||||||  |||||||||||||||||||
-Sbjct  365   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCCGGCCGGCAAATCCAGTTCTGCTCCTCTCCA  424
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| ||||||||||| |||||||||||||||||||||||||||| |||| ||||||||||
-Sbjct  425   GGCCCCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTGACCTCCCGGCCC-ATGGATGTGG  483
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             ||||||||||||||||||||||||||||||||||||| |||
-Sbjct  484   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAAAAG  524
-
-
- Score = 224 bits (248),  Expect = 4e-54
- Identities = 133/139 (96%), Gaps = 0/139 (0%)
- Strand=Plus/Plus
-
-Query  79   AACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGC  138
-            |||||||||||||||||||||||||||||||||||||||||||||| |||||||| ||||
-Sbjct  1    AACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGGAGCCCAGCACAGC  60
-
-Query  139  CACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGC  198
-            ||||||||| ||||||||||||||||||||||||||||| ||||||||||| ||||||||
-Sbjct  61   CACCGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTACTGGCCCTGGCCACACTTTGC  120
-
-Query  199  ATCGCTGGCCAGGCAGGTG  217
-            ||| |||||||||||||||
-Sbjct  121  ATCACTGGCCAGGCAGGTG  139
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  173  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  232
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  233  ATCAATGGCTGGG  245
-
-
- Score = 69.8 bits (76),  Expect = 2e-07
- Identities = 41/43 (95%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||| || |||||||||||||||||||||||||||||||
-Sbjct  133  GCAGGTGCAAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  175
-
-
->XM_011928386.1 PREDICTED: Colobus angolensis palliatus bone gamma-carboxyglutamate 
-(gla) protein (BGLAP), mRNA
-Length=738
-
- Score = 435 bits (482),  Expect = 1e-117
- Identities = 265/281 (94%), Gaps = 6/281 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  443   GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  502
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  503   CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  562
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             |||||||||||||||||||||||||||||||||||||||  |||||||||||||||||||
-Sbjct  563   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAAATCCAGTTCTGCTCCTCTCCA  622
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| |||||||      |||||||| ||||||||||||||||||||||||||||||||||
-Sbjct  623   GGCCCCCTTCT------TTCCCCTTCCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  676
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             ||||||||||||||||||||||||||||||||||||| |||
-Sbjct  677   GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAAAAG  717
-
-
- Score = 318 bits (352),  Expect = 2e-82
- Identities = 202/218 (93%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            ||||||||||||||||||||||| ||||||||||||| ||  |||||||||||||| |||
-Sbjct  120  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGCCACACCCCTCCTCATCGCCGGG  179
-
-Query  61   -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-             |||||||||||| ||||||||||||||||||||||| ||||||||||||||||||||||
-Sbjct  180  ACACAGCCCAGAGCGTATAAACAGTGCTGGAGGCTGGTGGGGCAGGCCAGCTGAGTCCTG  239
-
-Query  120  AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-            ||||| |||||||| |||||||| |||||||||||||||||||||||||||||||||  |
-Sbjct  240  AGCAGGAGCCCAGCACAGCCACCAAGACACCATGAGAGCCCTCACACTCCTCGCCCTGCT  299
-
-Query  180  GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||| ||||| ||||| |||||||||||||||
-Sbjct  300  GGCCCTGGCCACACTTCGCATCACTGGCCAGGCAGGTG  337
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  371  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  430
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  431  ATCAATGGCTGGG  443
-
-
- Score = 64.4 bits (70),  Expect = 8e-06
- Identities = 40/43 (93%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||| || ||||| |||||||||||||||||||||||||
-Sbjct  331  GCAGGTGCAAAACCCAGTGGTGCAGAGTCCAGCAAAGGTGCAG  373
-
-
->XM_017852456.1 PREDICTED: Rhinopithecus bieti bone gamma-carboxyglutamate protein 
-(BGLAP), mRNA
-Length=721
-
- Score = 430 bits (476),  Expect = 5e-116
- Identities = 268/287 (93%), Gaps = 7/287 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  415   GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  474
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| ||||||||||||| |||||||||||
-Sbjct  475   CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCAGCGCTTCTACG  534
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             |||||||||||||||||||||||||||||||||||||||  |||||||||||||||||||
-Sbjct  535   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAAATCCAGTTCTGCTCCTCTCCA  594
-
-Query  1131  GGCACCCTTCTTTCCTCTTC------CCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGG  1184
-             ||| ||||||||||| ||||      |||||||||||||||||||||||| |||| ||||
-Sbjct  595   GGCCCCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTTGCCCTGACCTCCCGGCCC-ATGG  653
-
-Query  1185  ATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             ||||||||||||||||||||||||||||||||||||||||||| |||
-Sbjct  654   ATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAAAAG  700
-
-
- Score = 327 bits (362),  Expect = 4e-85
- Identities = 204/218 (94%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            ||||||||||||||||||||||| ||||||||||||| ||  |||||||||||||| |||
-Sbjct  92   GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGCCACACCCCTCCTCATCGCCGGG  151
-
-Query  61   -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-             |||||||||||| ||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  152  ACACAGCCCAGAGCGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  211
-
-Query  120  AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-            ||||| |||||||| ||||||||||||| ||||||||||||||||||||||||||||  |
-Sbjct  212  AGCAGGAGCCCAGCACAGCCACCGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCT  271
-
-Query  180  GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||| ||||||||||| |||||||||||||||
-Sbjct  272  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  309
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  343  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  402
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  403  ATCAATGGCTGGG  415
-
-
- Score = 69.8 bits (76),  Expect = 2e-07
- Identities = 41/43 (95%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||| || |||||||||||||||||||||||||||||||
-Sbjct  303  GCAGGTGCAAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  345
-
-
->XR_001881540.1 PREDICTED: Rhinopithecus bieti polyamine-modulated factor 1 (LOC108515517), 
-transcript variant X3, misc_RNA
-Length=1393
-
- Score = 430 bits (476),  Expect = 5e-116
- Identities = 268/287 (93%), Gaps = 7/287 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  1103  GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  1162
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| ||||||||||||| |||||||||||
-Sbjct  1163  CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCAGCGCTTCTACG  1222
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             |||||||||||||||||||||||||||||||||||||||  |||||||||||||||||||
-Sbjct  1223  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAAATCCAGTTCTGCTCCTCTCCA  1282
-
-Query  1131  GGCACCCTTCTTTCCTCTTC------CCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGG  1184
-             ||| ||||||||||| ||||      |||||||||||||||||||||||| |||| ||||
-Sbjct  1283  GGCCCCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTTGCCCTGACCTCCCGGCCC-ATGG  1341
-
-Query  1185  ATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             ||||||||||||||||||||||||||||||||||||||||||| |||
-Sbjct  1342  ATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAAAAG  1388
-
-
- Score = 327 bits (362),  Expect = 4e-85
- Identities = 204/218 (94%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            ||||||||||||||||||||||| ||||||||||||| ||  |||||||||||||| |||
-Sbjct  780  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGCCACACCCCTCCTCATCGCCGGG  839
-
-Query  61   -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-             |||||||||||| ||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  840  ACACAGCCCAGAGCGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  899
-
-Query  120  AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-            ||||| |||||||| ||||||||||||| ||||||||||||||||||||||||||||  |
-Sbjct  900  AGCAGGAGCCCAGCACAGCCACCGAGACGCCATGAGAGCCCTCACACTCCTCGCCCTGCT  959
-
-Query  180  GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||| ||||||||||| |||||||||||||||
-Sbjct  960  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  997
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1031  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1090
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1091  ATCAATGGCTGGG  1103
-
-
- Score = 69.8 bits (76),  Expect = 2e-07
- Identities = 41/43 (95%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             |||||||| || |||||||||||||||||||||||||||||||
-Sbjct  991   GCAGGTGCAAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1033
-
-
->XM_007976754.1 PREDICTED: Chlorocebus sabaeus bone gamma-carboxyglutamate (gla) 
-protein (BGLAP), mRNA
-Length=2011
-
- Score = 430 bits (476),  Expect = 5e-116
- Identities = 268/287 (93%), Gaps = 7/287 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  1714  GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  1773
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1774  CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1833
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             |||||||||||||||| |||| ||||||||||||||||| ||||||||||||||||||||
-Sbjct  1834  GCCCGGTCTAGGGTGTTGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCA  1893
-
-Query  1131  GGCACCCTTCTTTC------CTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGG  1184
-             ||| ||||||||||      | ||| ||||||||||||||||||||||||||||||||||
-Sbjct  1894  GGCCCCCTTCTTTCCCCTTGCCCTTGCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGG  1953
-
-Query  1185  ATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||| |||||||||||||||||||||||| |||
-Sbjct  1954  ATGTGGGGTCCCCATCAT-CCAGCTGCTCCCAAATAAACTCCAAAAG  1999
-
-
- Score = 327 bits (362),  Expect = 4e-85
- Identities = 204/218 (94%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||| | |||||||||||||||||||||||||||||| |||
-Sbjct  1391  GGCAGATTCCCCCTAGACCCGCCTGTACCATGGTCAGGCATGCCCCTCCTCATCGCCGGG  1450
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1451  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  1510
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             ||||| |||||| | ||||||||||||| ||||||||||||||||||| |||||| |  |
-Sbjct  1511  AGCAGGAGCCCAACACAGCCACCGAGACGCCATGAGAGCCCTCACACTGCTCGCCTTGCT  1570
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||| ||||||||||| |||||||||||||||
-Sbjct  1571  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  1608
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1642  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1701
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1702  ATCAATGGCTGGG  1714
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             ||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  1602  GCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1644
-
-
->XM_003892817.4 PREDICTED: Papio anubis bone gamma-carboxyglutamate protein (BGLAP), 
-mRNA
-Length=1999
-
- Score = 414 bits (458),  Expect = 4e-111
- Identities = 261/281 (93%), Gaps = 7/281 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| ||| |||||||||||||||||||
-Sbjct  1717  GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGAGAGGTGTGTGAGCTCAATC  1776
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1777  CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1836
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||| ||||||||||||||||| ||||||||||||||||||||
-Sbjct  1837  GCCCGGTCTAGGGTGTCGCTCCGCTGGCCTGGCCGGCAAACCCAGTTCTGCTCCTCTCCA  1896
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| |||||||      |||||||| ||||||||||||||||||||||||||||||||||
-Sbjct  1897  GGCCCCCTTCT------TTCCCCTTCCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1950
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             ||| |||||||| ||||||||||||||||||||| || |||
-Sbjct  1951  GGTGCCCATCAT-CCAGCTGCTCCCAAATAAACTTCAAAAG  1990
-
-
- Score = 322 bits (356),  Expect = 2e-83
- Identities = 203/218 (93%), Gaps = 1/218 (0%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             ||||||||||||||||||||||| |||||||||||||||| ||||||||||||| | |||
-Sbjct  1394  GGCAGATTCCCCCTAGACCCGCCTGCACCATGGTCAGGCACGCCCCTCCTCATCACCGGG  1453
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1454  ACACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  1513
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             ||||| |||||| | ||||||| ||||| ||||||||||| ||||||||||||||||  |
-Sbjct  1514  AGCAGGAGCCCAACACAGCCACTGAGACGCCATGAGAGCCTTCACACTCCTCGCCCTGCT  1573
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||| ||||||||||| |||||||||||||||
-Sbjct  1574  GGCCCTGGCCACACTTTGCATCACTGGCCAGGCAGGTG  1611
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1645  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1704
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1705  ATCAATGGCTGGG  1717
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             ||||||||||| |||||||||||||||||||||||||||||||
-Sbjct  1605  GCAGGTGCGAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  1647
-
-
->XM_012449840.2 PREDICTED: Aotus nancymaae bone gamma-carboxyglutamate protein 
-(BGLAP), mRNA
-Length=1280
-
- Score = 403 bits (446),  Expect = 7e-108
- Identities = 258/281 (92%), Gaps = 4/281 (1%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||| ||||| ||| |||||||||||||||||||||||||||||||||||||||||||| |
-Sbjct  983   GAGTCCCAGCCCCTTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAACC  1042
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||| ||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1043  CGGACTGTGATGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1102
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||| | ||||||||||||| ||  ||||||| ||||||||||||||
-Sbjct  1103  GCCCGGTCTAGGGTGTCACCCTGCTGGCCTGGCAGGTGACCCCAGCTCTGCTCCTCTCCA  1162
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| |||||||    |||||||||  |||||||||||||| |||||||||  ||||||||
-Sbjct  1163  GGCCCCCTTCT----TCTTCCCCTCCCCCTTGCCCTGACCACCCAGCCCTGGGGATGTGG  1218
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  1219  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1259
-
-
- Score = 244 bits (270),  Expect = 4e-60
- Identities = 187/219 (85%), Gaps = 12/219 (5%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            |||||||||||||||||||| |||||||| ||||||||| |||||||||||||| | |||
-Sbjct  669  GGCAGATTCCCCCTAGACCCACCCGCACCGTGGTCAGGCGTGCCCCTCCTCATCTCCGGG  728
-
-Query  61   -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGG-CAGGCCAGCTGAGTCCT  118
-             |   |||||||||||||||||||||||||||| ||  |||| ||||||||| ||| |||
-Sbjct  729  GCGTGGCCCAGAGGGTATAAACAGTGCTGGAGGGTGCTGGGGGCAGGCCAGCCGAGCCCT  788
-
-Query  119  GAGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTAT  178
-            ||||||||||   |||        ||||| |||||||| |||||||||||||| || |  
-Sbjct  789  GAGCAGCAGCT--GCG--------GAGACGCCATGAGAACCCTCACACTCCTCCCCGTGC  838
-
-Query  179  TGGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||||||||||||||||||||||||||||||||||
-Sbjct  839  TGGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  877
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  911  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  970
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  971  ATCAATGGCTGGG  983
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  871  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  913
-
-
->BC113432.1 Homo sapiens bone gamma-carboxyglutamate (gla) protein, mRNA 
-(cDNA clone MGC:141992 IMAGE:8322484), complete cds
- BC113434.1 Homo sapiens bone gamma-carboxyglutamate (gla) protein, mRNA 
-(cDNA clone MGC:141994 IMAGE:8322486), complete cds
-Length=465
-
- Score = 399 bits (442),  Expect = 8e-107
- Identities = 221/221 (100%), Gaps = 0/221 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  245   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  304
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  305   CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  364
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  365   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  424
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCT  1171
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  425   GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCT  465
-
-
- Score = 251 bits (278),  Expect = 3e-62
- Identities = 139/139 (100%), Gaps = 0/139 (0%)
- Strand=Plus/Plus
-
-Query  79   AACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGC  138
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1    AACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGC  60
-
-Query  139  CACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGC  198
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  61   CACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGC  120
-
-Query  199  ATCGCTGGCCAGGCAGGTG  217
-            |||||||||||||||||||
-Sbjct  121  ATCGCTGGCCAGGCAGGTG  139
-
-
- Score = 132 bits (146),  Expect = 2e-26
- Identities = 73/73 (100%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  173  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  232
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  233  ATCAATGGCTGGG  245
-
-
- Score = 78.8 bits (86),  Expect = 4e-10
- Identities = 43/43 (100%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||||||||||||||||||||||||||
-Sbjct  133  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  175
-
-
->XM_010348451.1 PREDICTED: Saimiri boliviensis boliviensis bone gamma-carboxyglutamate 
-(gla) protein (BGLAP), transcript variant X3, mRNA
-Length=1118
-
- Score = 392 bits (434),  Expect = 1e-104
- Identities = 260/289 (90%), Gaps = 6/289 (2%)
- Strand=Plus/Plus
-
-Query  943   GTGTCTCAGAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGA  1002
-             ||| ||  ||| ||||| ||| ||||||||||||||||||||||||||||||||||||||
-Sbjct  826   GTGGCTGGGAGTCCCAGCCCCTTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGA  885
-
-Query  1003  GCTCAATCCGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCG  1062
-             |||||| ||||||||||| ||| ||||||||||||||||||| |||||||||||||||||
-Sbjct  886   GCTCAACCCGGACTGTGATGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCG  945
-
-Query  1063  CTTCTACGGCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCT  1122
-             ||||||||||||||||||||||||||| |||||| || ||||||  ||||||| ||||||
-Sbjct  946   CTTCTACGGCCCGGTCTAGGGTGTCGCCCTGCTGTCCCGGCCGGTGACCCCAGCTCTGCT  1005
-
-Query  1123  CCTCTCCAGGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTAT  1182
-             ||||||||||| |||      | |||||||||| |||||||||||||| ||| |||||  
-Sbjct  1006  CCTCTCCAGGCCCCC------CTTCTTCCCCTTCCCCTTGCCCTGACCACCCGGCCCTGG  1059
-
-Query  1183  GGATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1060  GGATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1108
-
-
- Score = 210 bits (232),  Expect = 8e-50
- Identities = 177/218 (81%), Gaps = 17/218 (8%)
- Strand=Plus/Plus
-
-Query  1    GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-            |||||||||||||||||||| |||| ||  ||||||||| |||||| || |||| |||||
-Sbjct  527  GGCAGATTCCCCCTAGACCCACCCGTACTGTGGTCAGGCGTGCCCCCCCCCATCTCTGGG  586
-
-Query  61   -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-             ||  ||||||||||||||||||||||||||||      |||||||||||||||| ||||
-Sbjct  587  GCAAGGCCCAGAGGGTATAAACAGTGCTGGAGG------GGGCAGGCCAGCTGAGCCCTG  640
-
-Query  120  AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-            |||||| ||            ||||||| |||||||||||||||  |||||  || |  |
-Sbjct  641  AGCAGCTGCTG----------CCGAGACGCCATGAGAGCCCTCATGCTCCTTCCCGTGGT  690
-
-Query  180  GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            ||||||||| ||||||||||| ||||||||||||||||
-Sbjct  691  GGCCCTGGCTGCACTTTGCATTGCTGGCCAGGCAGGTG  728
-
-
- Score = 118 bits (130),  Expect = 4e-22
- Identities = 70/73 (96%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||| ||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  762  CAGCTTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  821
-
-Query  738  ATCAATGGCTGGG  750
-            |||| ||||||||
-Sbjct  822  ATCAGTGGCTGGG  834
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||||||||||||||| ||||||||||||||||||||||
-Sbjct  722  GCAGGTGCGAAGCCCAGCGGAGCAGAGTCCAGCAAAGGTGCAG  764
-
-
->XM_010348450.1 PREDICTED: Saimiri boliviensis boliviensis bone gamma-carboxyglutamate 
-(gla) protein (BGLAP), transcript variant X2, mRNA
-Length=1817
-
- Score = 392 bits (434),  Expect = 1e-104
- Identities = 260/289 (90%), Gaps = 6/289 (2%)
- Strand=Plus/Plus
-
-Query  943   GTGTCTCAGAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGA  1002
-             ||| ||  ||| ||||| ||| ||||||||||||||||||||||||||||||||||||||
-Sbjct  1525  GTGGCTGGGAGTCCCAGCCCCTTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGA  1584
-
-Query  1003  GCTCAATCCGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCG  1062
-             |||||| ||||||||||| ||| ||||||||||||||||||| |||||||||||||||||
-Sbjct  1585  GCTCAACCCGGACTGTGATGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCG  1644
-
-Query  1063  CTTCTACGGCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCT  1122
-             ||||||||||||||||||||||||||| |||||| || ||||||  ||||||| ||||||
-Sbjct  1645  CTTCTACGGCCCGGTCTAGGGTGTCGCCCTGCTGTCCCGGCCGGTGACCCCAGCTCTGCT  1704
-
-Query  1123  CCTCTCCAGGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTAT  1182
-             ||||||||||| |||      | |||||||||| |||||||||||||| ||| |||||  
-Sbjct  1705  CCTCTCCAGGCCCCC------CTTCTTCCCCTTCCCCTTGCCCTGACCACCCGGCCCTGG  1758
-
-Query  1183  GGATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1759  GGATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1807
-
-
- Score = 210 bits (232),  Expect = 8e-50
- Identities = 177/218 (81%), Gaps = 17/218 (8%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             |||||||||||||||||||| |||| ||  ||||||||| |||||| || |||| |||||
-Sbjct  1226  GGCAGATTCCCCCTAGACCCACCCGTACTGTGGTCAGGCGTGCCCCCCCCCATCTCTGGG  1285
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              ||  ||||||||||||||||||||||||||||      |||||||||||||||| ||||
-Sbjct  1286  GCAAGGCCCAGAGGGTATAAACAGTGCTGGAGG------GGGCAGGCCAGCTGAGCCCTG  1339
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             |||||| ||            ||||||| |||||||||||||||  |||||  || |  |
-Sbjct  1340  AGCAGCTGCTG----------CCGAGACGCCATGAGAGCCCTCATGCTCCTTCCCGTGGT  1389
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             ||||||||| ||||||||||| ||||||||||||||||
-Sbjct  1390  GGCCCTGGCTGCACTTTGCATTGCTGGCCAGGCAGGTG  1427
-
-
- Score = 118 bits (130),  Expect = 4e-22
- Identities = 70/73 (96%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||| ||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1461  CAGCTTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1520
-
-Query  738   ATCAATGGCTGGG  750
-             |||| ||||||||
-Sbjct  1521  ATCAGTGGCTGGG  1533
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             |||||||||||||||||||| ||||||||||||||||||||||
-Sbjct  1421  GCAGGTGCGAAGCCCAGCGGAGCAGAGTCCAGCAAAGGTGCAG  1463
-
-
->XM_003937835.2 PREDICTED: Saimiri boliviensis boliviensis bone gamma-carboxyglutamate 
-(gla) protein (BGLAP), transcript variant X1, mRNA
-Length=1923
-
- Score = 392 bits (434),  Expect = 1e-104
- Identities = 260/289 (90%), Gaps = 6/289 (2%)
- Strand=Plus/Plus
-
-Query  943   GTGTCTCAGAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGA  1002
-             ||| ||  ||| ||||| ||| ||||||||||||||||||||||||||||||||||||||
-Sbjct  1631  GTGGCTGGGAGTCCCAGCCCCTTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGA  1690
-
-Query  1003  GCTCAATCCGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCG  1062
-             |||||| ||||||||||| ||| ||||||||||||||||||| |||||||||||||||||
-Sbjct  1691  GCTCAACCCGGACTGTGATGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCG  1750
-
-Query  1063  CTTCTACGGCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCT  1122
-             ||||||||||||||||||||||||||| |||||| || ||||||  ||||||| ||||||
-Sbjct  1751  CTTCTACGGCCCGGTCTAGGGTGTCGCCCTGCTGTCCCGGCCGGTGACCCCAGCTCTGCT  1810
-
-Query  1123  CCTCTCCAGGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTAT  1182
-             ||||||||||| |||      | |||||||||| |||||||||||||| ||| |||||  
-Sbjct  1811  CCTCTCCAGGCCCCC------CTTCTTCCCCTTCCCCTTGCCCTGACCACCCGGCCCTGG  1864
-
-Query  1183  GGATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||||||||||
-Sbjct  1865  GGATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1913
-
-
- Score = 210 bits (232),  Expect = 8e-50
- Identities = 177/218 (81%), Gaps = 17/218 (8%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             |||||||||||||||||||| |||| ||  ||||||||| |||||| || |||| |||||
-Sbjct  1332  GGCAGATTCCCCCTAGACCCACCCGTACTGTGGTCAGGCGTGCCCCCCCCCATCTCTGGG  1391
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              ||  ||||||||||||||||||||||||||||      |||||||||||||||| ||||
-Sbjct  1392  GCAAGGCCCAGAGGGTATAAACAGTGCTGGAGG------GGGCAGGCCAGCTGAGCCCTG  1445
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             |||||| ||            ||||||| |||||||||||||||  |||||  || |  |
-Sbjct  1446  AGCAGCTGCTG----------CCGAGACGCCATGAGAGCCCTCATGCTCCTTCCCGTGGT  1495
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             ||||||||| ||||||||||| ||||||||||||||||
-Sbjct  1496  GGCCCTGGCTGCACTTTGCATTGCTGGCCAGGCAGGTG  1533
-
-
- Score = 118 bits (130),  Expect = 4e-22
- Identities = 70/73 (96%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||| ||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1567  CAGCTTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1626
-
-Query  738   ATCAATGGCTGGG  750
-             |||| ||||||||
-Sbjct  1627  ATCAGTGGCTGGG  1639
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             |||||||||||||||||||| ||||||||||||||||||||||
-Sbjct  1527  GCAGGTGCGAAGCCCAGCGGAGCAGAGTCCAGCAAAGGTGCAG  1569
-
-
->XM_008984594.2 PREDICTED: Callithrix jacchus bone gamma-carboxyglutamate protein 
-(BGLAP), mRNA
-Length=1744
-
- Score = 383 bits (424),  Expect = 6e-102
- Identities = 253/281 (90%), Gaps = 7/281 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||| ||||  ||| |||||||||||||||||||||||||||||||||||||||||||| |
-Sbjct  1236  GAGTCCCACCCCCTTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAACC  1295
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||| ||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1296  CGGACTGTGATGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1355
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||| ||||||||| ||||||  ||||||| ||||||||||||||
-Sbjct  1356  GCCCGGTCTAGGGTGTCGCCCTGCTGGCCCGGCCGGTGACCCCAGCTCTGCTCCTCTCCA  1415
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| ||||       |||||||||  |||||||||||||| ||  |||||  ||||||||
-Sbjct  1416  GGCCCCCT-------TCTTCCCCTCCCCCTTGCCCTGACCCCCTGGCCCTGGGGATGTGG  1468
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  1469  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1509
-
-
- Score = 248 bits (274),  Expect = 3e-61
- Identities = 187/219 (85%), Gaps = 12/219 (5%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCT-GG  59
-             |||||||||||||||||||| |||| ||||||||||||| |||||||||||||| |  ||
-Sbjct  922   GGCAGATTCCCCCTAGACCCACCCGAACCATGGTCAGGCGTGCCCCTCCTCATCTCCAGG  981
-
-Query  60    GCACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGG-CAGGCCAGCTGAGTCCT  118
-             |||  |||||||||||||||||||||||||||| || ||||| ||||||||| ||| |||
-Sbjct  982   GCATGGCCCAGAGGGTATAAACAGTGCTGGAGGGTGCCGGGGACAGGCCAGCCGAGCCCT  1041
-
-Query  119   GAGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTAT  178
-             ||||||||||            ||||| | ||||||||||||||||||||||| || |  
-Sbjct  1042  GAGCAGCAGCTG----------CCGAGCCGCCATGAGAGCCCTCACACTCCTCCCCGTGC  1091
-
-Query  179   TGGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             |||||||||||||||||||||| ||||||||||||||||
-Sbjct  1092  TGGCCCTGGCCGCACTTTGCATTGCTGGCCAGGCAGGTG  1130
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1164  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1223
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1224  ATCAATGGCTGGG  1236
-
-
- Score = 73.4 bits (80),  Expect = 1e-08
- Identities = 42/43 (98%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             |||||||||||||||||| ||||||||||||||||||||||||
-Sbjct  1124  GCAGGTGCGAAGCCCAGCAGTGCAGAGTCCAGCAAAGGTGCAG  1166
-
-
->XM_017510005.1 PREDICTED: Cebus capucinus imitator bone gamma-carboxyglutamate 
-protein (BGLAP), mRNA
-Length=1860
-
- Score = 369 bits (408),  Expect = 1e-97
- Identities = 250/281 (89%), Gaps = 7/281 (2%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||| ||||| ||| |||||||||||||||||||||||||||||||||||||||||||| |
-Sbjct  1580  GAGTCCCAGCCCCTTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAACC  1639
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||| ||| ||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1640  CGGACTGTGATGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCTTCTACG  1699
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             |||| |||||| |||| || |||||||||  |||||  ||||||| ||||||||||||||
-Sbjct  1700  GCCCAGTCTAGAGTGTTGCCCTGCTGGCCCAGCCGGTGACCCCAGCTCTGCTCCTCTCCA  1759
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTATGGATGTGG  1190
-             ||| ||||       |||||||||  |||||||||||||| ||| |||||   |||||||
-Sbjct  1760  GGCCCCCT-------TCTTCCCCTCCCCCTTGCCCTGACCACCCGGCCCTGGAGATGTGG  1812
-
-Query  1191  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |||||||||||||||||||||||||||||||||||||||||
-Sbjct  1813  GGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1853
-
-
- Score = 230 bits (254),  Expect = 9e-56
- Identities = 181/218 (83%), Gaps = 19/218 (9%)
- Strand=Plus/Plus
-
-Query  1     GGCAGATTCCCCCTAGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGG  60
-             |||||||||||||||||||| ||| |||||||||||||| |||||||||||||| | |||
-Sbjct  1275  GGCAGATTCCCCCTAGACCCACCCACACCATGGTCAGGCGTGCCCCTCCTCATCTCCGGG  1334
-
-Query  61    -CACAGCCCAGAGGGTATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTG  119
-              ||  |||||||||||||||||||||||||||       |||||||||||| ||| ||||
-Sbjct  1335  GCATGGCCCAGAGGGTATAAACAGTGCTGGAGA------GGGCAGGCCAGCCGAGCCCTG  1388
-
-Query  120   AGCAGCAGCCCAGCGCAGCCACCGAGACACCATGAGAGCCCTCACACTCCTCGCCCTATT  179
-             |||||||||              ||||| ||||||||||||||||||||||| ||    |
-Sbjct  1389  AGCAGCAGCTG------------GAGACGCCATGAGAGCCCTCACACTCCTCCCCGCGCT  1436
-
-Query  180   GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-             ||||||||||||||||||||||||||||||||||||||
-Sbjct  1437  GGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTG  1474
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678   CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-             |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  1508  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  1567
-
-Query  738   ATCAATGGCTGGG  750
-             |||||||||||||
-Sbjct  1568  ATCAATGGCTGGG  1580
-
-
- Score = 60.8 bits (66),  Expect = 9e-05
- Identities = 39/43 (91%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468   GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-             |||||||| ||||||||||| |  |||||||||||||||||||
-Sbjct  1468  GCAGGTGCAAAGCCCAGCGGAGTGGAGTCCAGCAAAGGTGCAG  1510
-
-
->XM_010366603.1 PREDICTED: Rhinopithecus roxellana osteocalcin (LOC104664918), 
-mRNA
-Length=392
-
- Score = 343 bits (380),  Expect = 6e-90
- Identities = 207/218 (95%), Gaps = 0/218 (0%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCCGGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGCTCAATC  1010
-             ||||||||| ||||||||| |||||||||||||||| |||||||||||||||||||||||
-Sbjct  175   GAGCCCCAGCCCCCTACCCAGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGCTCAATC  234
-
-Query  1011  CGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCTTCTACG  1070
-             |||||||||||||| ||||||||||||||||||| ||||||||||||| |||||||||||
-Sbjct  235   CGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCAGCGCTTCTACG  294
-
-Query  1071  GCCCGGTCTAGGGTGTCGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCTGCTCCTCTCCA  1130
-             ||||||||||||||||||||||||||||| |||||||||  |||||||||||||||||||
-Sbjct  295   GCCCGGTCTAGGGTGTCGCTCTGCTGGCCCGGCCGGCAAATCCAGTTCTGCTCCTCTCCA  354
-
-Query  1131  GGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGA  1168
-             ||| ||||||||||| ||||||||||||||||||||||
-Sbjct  355   GGCCCCCTTCTTTCCCCTTCCCCTTGCCCTTGCCCTGA  392
-
-
- Score = 127 bits (140),  Expect = 8e-25
- Identities = 72/73 (99%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||||||||||||||||||||||||||||| |||||||||||||||||||||||||
-Sbjct  103  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTGGTGAAGAGACCCAGGCGCTACCTGT  162
-
-Query  738  ATCAATGGCTGGG  750
-            |||||||||||||
-Sbjct  163  ATCAATGGCTGGG  175
-
-
- Score = 107 bits (118),  Expect = 7e-19
- Identities = 65/69 (94%), Gaps = 0/69 (0%)
- Strand=Plus/Plus
-
-Query  149  CCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGCATCGCTGGCC  208
-            ||||||||||||||||||||||||||||  ||||||||||| ||||||||||| ||||||
-Sbjct  1    CCATGAGAGCCCTCACACTCCTCGCCCTGCTGGCCCTGGCCACACTTTGCATCACTGGCC  60
-
-Query  209  AGGCAGGTG  217
-            |||||||||
-Sbjct  61   AGGCAGGTG  69
-
-
- Score = 69.8 bits (76),  Expect = 2e-07
- Identities = 41/43 (95%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||||||| || |||||||||||||||||||||||||||||||
-Sbjct  63   GCAGGTGCAAAACCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  105
-
-
->XM_012658608.1 PREDICTED: Propithecus coquereli bone gamma-carboxyglutamate 
-(gla) protein (BGLAP), mRNA
-Length=601
-
- Score = 293 bits (324),  Expect = 9e-75
- Identities = 241/289 (83%), Gaps = 21/289 (7%)
- Strand=Plus/Plus
-
-Query  951   GAGCCCCAGTCCCCTACCC------GGATCCCCTGGAGCCCAGGAGGGAGGTGTGTGAGC  1004
-             ||||||||| |||||||||      ||||||||||||||||| |||||||||||||||||
-Sbjct  321   GAGCCCCAGCCCCCTACCCATACCTGGATCCCCTGGAGCCCAAGAGGGAGGTGTGTGAGC  380
-
-Query  1005  TCAATCCGGACTGTGACGAGTTGGCTGACCACATCGGCTTTCAGGAGGCCTATCGGCGCT  1064
-             |||| ||||||||||||||| ||||||||||||||||||| |||||||||||||||||||
-Sbjct  381   TCAACCCGGACTGTGACGAGCTGGCTGACCACATCGGCTTCCAGGAGGCCTATCGGCGCT  440
-
-Query  1065  TCTACGGCCCGGTCTAGGGTGT-CGCTCTGCTGGCCTGGCCGGCAACCCCAGTTCT-GCT  1122
-             | |||||| ||| ||||||  | ||  || ||||||  || ||||| ||||| ||| |||
-Sbjct  441   TTTACGGCACGGCCTAGGGCATCCGTCCTCCTGGCCCAGCTGGCAAGCCCAGCTCTGGCT  500
-
-Query  1123  CCTCTCCAGGCACCCTTCTTTCCTCTTCCCCTTGCCCTTGCCCTGACCTCCCAGCCCTAT  1182
-             ||||     ||||       ||| ||| ||||||||||  |||||||||||| |||||||
-Sbjct  501   CCTC-----GCAC------ATCC-CTT-CCCTTGCCCTCACCCTGACCTCCCTGCCCTAT  547
-
-Query  1183  GGATGTGGGGTCCCCATCATCCCAGCTGCTCCCAAATAAACTCCAGAAG  1231
-             |   ||||||||||||||||||||||||||||  |||||||||||||||
-Sbjct  548   GAGGGTGGGGTCCCCATCATCCCAGCTGCTCCAGAATAAACTCCAGAAG  596
-
-
- Score = 197 bits (218),  Expect = 5e-46
- Identities = 174/211 (82%), Gaps = 11/211 (5%)
- Strand=Plus/Plus
-
-Query  16   GACCCGCCCGCACCATGGTCAGGCATGCCCCT--CCTCATCGCT-GGGCACAGCCCAGAG  72
-            ||||||||  || ||||||||||  |||||||  || || | |  ||||||||||  |||
-Sbjct  7    GACCCGCCTACAACATGGTCAGG--TGCCCCTGCCCCCACCTCCAGGGCACAGCCTGGAG  64
-
-Query  73   GGTATAAACAGTGCTGGAGGCTGGC-GGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCA  131
-            ||||||||||||||||||||||||  ||||||| ||||||||||||||  ||||||||| 
-Sbjct  65   GGTATAAACAGTGCTGGAGGCTGGAAGGGGCAGACCAGCTGAGTCCTGGACAGCAGCCCT  124
-
-Query  132  G-CGCAGCCACCGAGA----CACCATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTG  186
-              ||||||||| || |    ||||||||| ||||||||||||||||| ||  ||||||||
-Sbjct  125  CACGCAGCCACTGACAGACGCACCATGAGGGCCCTCACACTCCTCGCTCTGCTGGCCCTG  184
-
-Query  187  GCCGCACTTTGCATCGCTGGCCAGGCAGGTG  217
-            |||||||| ||| ||||||| | ||||| ||
-Sbjct  185  GCCGCACTGTGCCTCGCTGGACTGGCAGATG  215
-
-
- Score = 96.9 bits (106),  Expect = 1e-15
- Identities = 65/73 (89%), Gaps = 0/73 (0%)
- Strand=Plus/Plus
-
-Query  678  CAGCCTTTGTGTCCAAGCAGGAGGGCAGCGAGGTAGTGAAGAGACCCAGGCGCTACCTGT  737
-            |||||||  ||||||||||||||||||| |||||  | |||||||||| |||||||||||
-Sbjct  249  CAGCCTTCATGTCCAAGCAGGAGGGCAGTGAGGTGATAAAGAGACCCAAGCGCTACCTGT  308
-
-Query  738  ATCAATGGCTGGG  750
-            |||| ||||||||
-Sbjct  309  ATCACTGGCTGGG  321
-
-
- Score = 51.8 bits (56),  Expect = 0.049
- Identities = 37/43 (86%), Gaps = 0/43 (0%)
- Strand=Plus/Plus
-
-Query  468  GCAGGTGCGAAGCCCAGCGGTGCAGAGTCCAGCAAAGGTGCAG  510
-            |||| ||||||||||||| ||||||||||   || ||||||||
-Sbjct  209  GCAGATGCGAAGCCCAGCAGTGCAGAGTCTGACAGAGGTGCAG  251
-
-
-  Database: Nucleotide collection (nt)
-    Posted date:  Oct 13, 2017  4:35 PM
-  Number of letters in database: 160,111,102,915
-  Number of sequences in database:  44,928,843
-
-Lambda      K        H
-   0.634    0.408    0.912 
-Gapped
-Lambda      K        H
-   0.625    0.410    0.780 
-Matrix: blastn matrix:2 -3
-Gap Penalties: Existence: 5, Extension: 2
-Number of Sequences: 44928843
-Number of Hits to DB: 26317416
-Number of extensions: 51124
-Number of successful extensions: 51124
-Number of sequences better than 10: 7
-Number of HSP's better than 10 without gapping: 0
-Number of HSP's gapped: 51124
-Number of HSP's successfully gapped: 14
-Length of query: 1231
-Length of database: 160111102915
-Length adjustment: 38
-Effective length of query: 1193
-Effective length of database: 158403806881
-Effective search space: 188975741609033
-Effective search space used: 188975741609033
-A: 0
-X1: 22 (20.1 bits)
-X2: 33 (29.8 bits)
-X3: 110 (99.2 bits)
-S1: 28 (26.5 bits)
-S2: 48 (44.6 bits)
-
-
-
-`;
-
-    a = 'BLASTN 2.7.0+\nReference: Stephen F. Altschul, Thomas L. Madden, Alejandro\nA. Schaffer, Jinghui Zhang, Zheng Zhang, Webb Miller, and\nDavid J. Lipman (1997), "Gapped BLAST and PSI-BLAST: a new\ngeneration of protein database search programs", Nucleic\nAcids Res. 25:3389-3402.\n\n\nRID: YP2X522301R\n\n\nDatabase: Nucleotide collection (nt)\n           44,954,477 sequences; 160,206,672,696 total letters\n\nQuery= \nLength=2\n\n\n<b>No significant similarity found.</b> For reasons why, <A HREF = "Blast.cgi?CMD=Web&PAGE_TYPE=BlastDocs&DOC_TYPE=FAQ"><b>click here</A>.</b><br><br>\n  Database: Nucleotide collection (nt)\n    Posted date:  Oct 18, 2017  11:25 PM\n  Number of letters in database: 160,206,672,696\n  Number of sequences in database:  44,954,477\n\nLambda      K        H\n   0.634    0.408    0.912 \nGapped\nLambda      K        H\n   0.625    0.410    0.780 \nMatrix: blastn matrix:2 -3\nGap Penalties: Existence: 5, Extension: 2\nNumber of Sequences: 44954477\nNumber of Hits to DB: 0\nNumber of extensions: 0\nNumber of successful extensions: 0\nNumber of sequences better than 10: 0\nNumber of HSP\'s better than 10 without gapping: 0\nNumber of HSP\'s gapped: 0\nNumber of HSP\'s successfully gapped: 0\nLength of query: 2\nLength of database: 160206672696\nLength adjustment: 0\nEffective length of query: 2\nEffective length of database: 160206672696\nEffective search space: 320413345392\nEffective search space used: 320413345392\nA: 0\nX1: 22 (20.1 bits)\nX2: 33 (29.8 bits)\nX3: 110 (99.2 bits)\nS1: 28 (26.5 bits)\nS2: 38 (35.6 bits)\n\n\n\n\n';
-
     execute(...taskParams) {
-        // console.log(Blast.data === this.a);
-        // const tempPath = path.join(__dirname, uuidv4());
-
-        // fs.writeFileSync(tempPath, taskParams);
-
         const wrap = new WrapBlast('blastn', 'nr', taskParams[0]);
-        //
         return wrap.exec();
-        //
-        // return new Promise(resolve => resolve(Blast.data))
     }
 }
