+++ bioinformatics-software-framework/src/app/components/drawing-board/drawing-board.component.css
