+++ bioinformatics-software-framework/src/app/components/drawing-board/drawing-board.component.spec.ts
