+++ bioinformatics-software-framework/src/app/components/drawing-board/step-box/step-box.component.html
