+++ bioinformatics-software-framework/src/app/pages/home-page/home-page.component.html
